@@ -1,9 +1,11 @@
 const homepage = 'https://mifi.no/losslesscut/';
 const githubLink = 'https://github.com/mifi/lossless-cut/';
-const releasesPage = 'https://github.com/mifi/lossless-cut/releases';
+const getReleaseUrl = (version) => `https://github.com/mifi/lossless-cut/releases/tag/v${version}`;
+const licensesPage = 'https://losslesscut.mifi.no/licenses.txt';
 
 module.exports = {
   homepage,
-  releasesPage,
+  getReleaseUrl,
   githubLink,
+  licensesPage,
 };
