@@ -1,88 +1,101 @@
-const electron = require('electron'); // eslint-disable-line
-const i18n = require('i18next');
+const electron = require('electron');
+const { t } = require('i18next');
+
+// menu-safe i18n.t:
+// https://github.com/mifi/lossless-cut/issues/1456
+const esc = (val) => val.replace(/&/g, '&&');
 
 const { Menu } = electron;
-const { dialog } = electron;
 
-const { homepage, releasesPage } = require('./constants');
+const { homepage, getReleaseUrl, licensesPage } = require('./constants');
 
-module.exports = (app, mainWindow, newVersion) => {
+module.exports = ({ app, mainWindow, newVersion, isStoreBuild }) => {
   const menu = [
     ...(process.platform === 'darwin' ? [{ role: 'appMenu' }] : []),
 
     {
-      label: i18n.t('File'),
+      label: esc(t('File')),
       submenu: [
         {
-          label: i18n.t('Open'),
+          label: esc(t('Open')),
           accelerator: 'CmdOrCtrl+O',
           async click() {
-            const { canceled, filePaths } = await dialog.showOpenDialog({ properties: ['openFile', 'multiSelections'] });
-            if (canceled) return;
-            mainWindow.webContents.send('file-opened', filePaths);
+            mainWindow.webContents.send('openFilesDialog');
           },
         },
         {
-          label: i18n.t('Close'),
+          label: esc(t('Close')),
           accelerator: 'CmdOrCtrl+W',
           async click() {
-            mainWindow.webContents.send('close-file');
+            mainWindow.webContents.send('closeCurrentFile');
           },
         },
         {
-          label: i18n.t('Close batch'),
+          label: esc(t('Close batch')),
           async click() {
-            mainWindow.webContents.send('close-batch-files');
+            mainWindow.webContents.send('closeBatchFiles');
           },
         },
         { type: 'separator' },
         {
-          label: i18n.t('Import project (LLC)...'),
+          label: esc(t('Import project (LLC)...')),
           click() {
             mainWindow.webContents.send('importEdlFile', 'llc');
           },
         },
         {
-          label: i18n.t('Export project (LLC)...'),
+          label: esc(t('Export project (LLC)...')),
           click() {
             mainWindow.webContents.send('exportEdlFile', 'llc');
           },
         },
         {
-          label: i18n.t('Import project'),
+          label: esc(t('Import project')),
           submenu: [
             {
-              label: i18n.t('LosslessCut (CSV)'),
+              label: esc(t('Times in seconds (CSV)')),
               click() {
                 mainWindow.webContents.send('importEdlFile', 'csv');
               },
             },
             {
-              label: i18n.t('EDL (MPlayer)'),
+              label: esc(t('Frame numbers (CSV)')),
+              click() {
+                mainWindow.webContents.send('importEdlFile', 'csv-frames');
+              },
+            },
+            {
+              label: esc(t('EDL (MPlayer)')),
               click() {
                 mainWindow.webContents.send('importEdlFile', 'mplayer');
               },
             },
             {
-              label: i18n.t('Text chapters / YouTube'),
+              label: esc(t('Text chapters / YouTube')),
               click() {
                 mainWindow.webContents.send('importEdlFile', 'youtube');
               },
             },
             {
-              label: i18n.t('DaVinci Resolve / Final Cut Pro XML'),
+              label: esc(t('DaVinci Resolve / Final Cut Pro XML')),
               click() {
                 mainWindow.webContents.send('importEdlFile', 'xmeml');
               },
             },
             {
-              label: i18n.t('CUE sheet file'),
+              label: esc(t('Final Cut Pro FCPX / FCPXML')),
+              click() {
+                mainWindow.webContents.send('importEdlFile', 'fcpxml');
+              },
+            },
+            {
+              label: esc(t('CUE sheet file')),
               click() {
                 mainWindow.webContents.send('importEdlFile', 'cue');
               },
             },
             {
-              label: i18n.t('PotPlayer Bookmarks (.pbf)'),
+              label: esc(t('PotPlayer Bookmarks (.pbf)')),
               click() {
                 mainWindow.webContents.send('importEdlFile', 'pbf');
               },
@@ -90,28 +103,34 @@ module.exports = (app, mainWindow, newVersion) => {
           ],
         },
         {
-          label: i18n.t('Export project'),
+          label: esc(t('Export project')),
           submenu: [
             {
-              label: i18n.t('LosslessCut (CSV)'),
+              label: esc(t('Times in seconds (CSV)')),
               click() {
                 mainWindow.webContents.send('exportEdlFile', 'csv');
               },
             },
             {
-              label: i18n.t('Timestamps (CSV)'),
+              label: esc(t('Timestamps (CSV)')),
               click() {
                 mainWindow.webContents.send('exportEdlFile', 'csv-human');
               },
             },
             {
-              label: i18n.t('Timestamps (TSV/TXT)'),
+              label: esc(t('Frame numbers (CSV)')),
+              click() {
+                mainWindow.webContents.send('exportEdlFile', 'csv-frames');
+              },
+            },
+            {
+              label: esc(t('Timestamps (TSV/TXT)')),
               click() {
                 mainWindow.webContents.send('exportEdlFile', 'tsv-human');
               },
             },
             {
-              label: i18n.t('Start times as YouTube Chapters'),
+              label: esc(t('Start times as YouTube Chapters')),
               click() {
                 mainWindow.webContents.send('exportEdlYouTube');
               },
@@ -120,13 +139,13 @@ module.exports = (app, mainWindow, newVersion) => {
         },
         { type: 'separator' },
         {
-          label: i18n.t('Convert to supported format'),
+          label: esc(t('Convert to supported format')),
           click() {
             mainWindow.webContents.send('html5ify');
           },
         },
         {
-          label: i18n.t('Fix incorrect duration'),
+          label: esc(t('Fix incorrect duration')),
           click() {
             mainWindow.webContents.send('fixInvalidDuration');
           },
@@ -135,79 +154,127 @@ module.exports = (app, mainWindow, newVersion) => {
 
         { type: 'separator' },
         {
-          label: i18n.t('Settings'),
+          label: esc(t('Settings')),
           accelerator: 'CmdOrCtrl+,',
           click() {
-            mainWindow.webContents.send('openSettings');
+            mainWindow.webContents.send('toggleSettings');
           },
         },
-        { type: 'separator' },
-        {
-          label: i18n.t('Exit'),
-          click() {
-            app.quit();
+        // Due to Apple Review Guidelines, we cannot include an Exit menu item here
+        // Apple has their own Quit from the app menu
+        ...(process.platform !== 'darwin' ? [
+          { type: 'separator' },
+          {
+            label: esc(t('Exit')),
+            click() {
+              app.quit();
+            },
           },
-        },
+        ] : []),
       ],
     },
 
     {
-      label: i18n.t('Edit'),
+      label: esc(t('Edit')),
       submenu: [
-        { role: 'undo', label: i18n.t('Undo') },
-        { role: 'redo', label: i18n.t('Redo') },
+        // https://github.com/mifi/lossless-cut/issues/610
+        // https://github.com/mifi/lossless-cut/issues/1183
+        { role: 'undo', label: esc(t('Undo')) },
+        { role: 'redo', label: esc(t('Redo')) },
         { type: 'separator' },
-        { role: 'cut', label: i18n.t('Cut') },
-        { role: 'copy', label: i18n.t('Copy') },
-        { role: 'paste', label: i18n.t('Paste') },
-        { role: 'selectall', label: i18n.t('Select All') },
+        { role: 'cut', label: esc(t('Cut')) },
+        { role: 'copy', label: esc(t('Copy')) },
+        { role: 'paste', label: esc(t('Paste')) },
+        { role: 'selectall', label: esc(t('Select All')) },
         { type: 'separator' },
         {
-          label: i18n.t('Segments'),
+          label: esc(t('Segments')),
           submenu: [
             {
-              label: i18n.t('Clear all segments'),
+              label: esc(t('Clear all segments')),
               click() {
                 mainWindow.webContents.send('clearSegments');
               },
             },
             {
-              label: i18n.t('Reorder segments by start time'),
+              label: esc(t('Reorder segments by start time')),
               click() {
                 mainWindow.webContents.send('reorderSegsByStartTime');
               },
             },
             {
-              label: i18n.t('Create num segments'),
+              label: esc(t('Create num segments')),
               click() {
                 mainWindow.webContents.send('createNumSegments');
               },
             },
             {
-              label: i18n.t('Create fixed duration segments'),
+              label: esc(t('Create fixed duration segments')),
               click() {
                 mainWindow.webContents.send('createFixedDurationSegments');
               },
             },
             {
-              label: i18n.t('Invert all segments on timeline'),
+              label: esc(t('Create random segments')),
               click() {
-                mainWindow.webContents.send('invertAllCutSegments');
+                mainWindow.webContents.send('createRandomSegments');
+              },
+            },
+            {
+              label: esc(t('Invert all segments on timeline')),
+              click() {
+                mainWindow.webContents.send('invertAllSegments');
+              },
+            },
+            {
+              label: esc(t('Fill gaps between segments')),
+              click() {
+                mainWindow.webContents.send('fillSegmentsGaps');
+              },
+            },
+            {
+              label: esc(t('Combine overlapping segments')),
+              click() {
+                mainWindow.webContents.send('combineOverlappingSegments');
+              },
+            },
+            {
+              label: esc(t('Combine selected segments')),
+              click() {
+                mainWindow.webContents.send('combineSelectedSegments');
+              },
+            },
+            {
+              label: esc(t('Shuffle segments order')),
+              click() {
+                mainWindow.webContents.send('shuffleSegments');
+              },
+            },
+            {
+              label: esc(t('Shift all segments on timeline')),
+              click() {
+                mainWindow.webContents.send('shiftAllSegmentTimes');
+              },
+            },
+            {
+              label: esc(t('Align segment times to keyframes')),
+              click() {
+                mainWindow.webContents.send('alignSegmentTimesToKeyframes');
               },
             },
           ],
         },
         {
-          label: i18n.t('Tracks'),
+          label: esc(t('Tracks')),
           submenu: [
             {
-              label: i18n.t('Extract all tracks'),
+              label: esc(t('Extract all tracks')),
               click() {
-                mainWindow.webContents.send('extract-all-streams');
+                mainWindow.webContents.send('extractAllStreams');
               },
             },
             {
-              label: i18n.t('Edit tracks / metadata tags'),
+              label: esc(t('Edit tracks / metadata tags')),
               click() {
                 mainWindow.webContents.send('showStreamsSelector');
               },
@@ -218,80 +285,115 @@ module.exports = (app, mainWindow, newVersion) => {
     },
 
     {
-      label: i18n.t('View'),
+      label: esc(t('View')),
       submenu: [
-        { role: 'togglefullscreen', label: i18n.t('Toggle Full Screen') },
+        { role: 'togglefullscreen', label: esc(t('Toggle Full Screen')) },
       ],
     },
 
     // On Windows the windowMenu has a close Ctrl+W which clashes with File->Close shortcut
     ...(process.platform === 'darwin'
-      ? [{ role: 'windowMenu', label: i18n.t('Window') }]
+      ? [{ role: 'windowMenu', label: esc(t('Window')) }]
       : [{
-        label: i18n.t('Window'),
-        submenu: [{ role: 'minimize', label: i18n.t('Minimize') }],
+        label: esc(t('Window')),
+        submenu: [{ role: 'minimize', label: esc(t('Minimize')) }],
       }]
     ),
 
     {
-      label: i18n.t('Tools'),
+      label: esc(t('Tools')),
       submenu: [
         {
-          label: i18n.t('Merge files'),
+          label: esc(t('Merge/concatenate files')),
           click() {
-            mainWindow.webContents.send('show-merge-dialog', true);
+            mainWindow.webContents.send('concatCurrentBatch');
           },
         },
         {
-          label: i18n.t('Batch convert to supported format'),
+          label: esc(t('Set custom start offset/timecode')),
           click() {
-            mainWindow.webContents.send('batchConvertFriendlyFormat');
+            mainWindow.webContents.send('askSetStartTimeOffset');
           },
         },
         {
-          label: i18n.t('Set custom start offset/timecode'),
+          label: esc(t('Detect black scenes')),
           click() {
-            mainWindow.webContents.send('set-start-offset', true);
+            mainWindow.webContents.send('detectBlackScenes');
           },
         },
-        { role: 'toggleDevTools', label: i18n.t('Toggle Developer Tools') },
+        {
+          label: esc(t('Detect silent scenes')),
+          click() {
+            mainWindow.webContents.send('detectSilentScenes');
+          },
+        },
+        {
+          label: esc(t('Detect scene changes')),
+          click() {
+            mainWindow.webContents.send('detectSceneChanges');
+          },
+        },
+        {
+          label: esc(t('Create segments from keyframes')),
+          click() {
+            mainWindow.webContents.send('createSegmentsFromKeyframes');
+          },
+        },
+        {
+          label: esc(t('Last ffmpeg commands')),
+          click() { mainWindow.webContents.send('toggleLastCommands'); },
+        },
+        { type: 'separator' },
+        { role: 'toggleDevTools', label: esc(t('Toggle Developer Tools')) },
       ],
     },
     {
       role: 'help',
-      label: i18n.t('Help'),
+      label: esc(t('Help')),
       submenu: [
         {
-          label: i18n.t('Help and shortcuts'),
-          click() {
-            mainWindow.webContents.send('openHelp');
-          },
+          label: esc(t('How to use')),
+          click() { electron.shell.openExternal('https://mifi.no/losslesscut/usage'); },
         },
         {
-          label: i18n.t('About'),
-          click() {
-            mainWindow.webContents.send('openAbout');
-          },
+          label: esc(t('FAQ')),
+          click() { electron.shell.openExternal('https://mifi.no/losslesscut/faq'); },
         },
         {
-          label: i18n.t('Learn More'),
+          label: esc(t('Troubleshooting')),
+          click() { electron.shell.openExternal('https://mifi.no/losslesscut/troubleshooting'); },
+        },
+        {
+          label: esc(t('Learn More')),
           click() { electron.shell.openExternal(homepage); },
         },
         {
-          label: i18n.t('Report an error'),
+          label: esc(t('Licenses')),
+          click() { electron.shell.openExternal(licensesPage); },
+        },
+        { type: 'separator' },
+        {
+          label: esc(t('Keyboard & mouse shortcuts')),
+          click() {
+            mainWindow.webContents.send('toggleKeyboardShortcuts');
+          },
+        },
+        {
+          label: esc(t('Report an error')),
           click() { mainWindow.webContents.send('openSendReportDialog'); },
         },
+        ...(process.platform !== 'darwin' ? [{ role: 'about', label: esc(t('About LosslessCut')) }] : []),
       ],
     },
   ];
 
-  if (newVersion) {
+  if (!isStoreBuild && newVersion) {
     menu.push({
-      label: i18n.t('New version!'),
+      label: esc(t('New version!')),
       submenu: [
         {
-          label: i18n.t('Download {{version}}', { version: newVersion }),
-          click() { electron.shell.openExternal(releasesPage); },
+          label: esc(t('Download {{version}}', { version: newVersion })),
+          click() { electron.shell.openExternal(getReleaseUrl(newVersion)); },
         },
       ],
     });
