@@ -1,58 +1,72 @@
-import React, { memo, useMemo, useRef } from 'react';
-import prettyMs from 'pretty-ms';
-import { FaSave, FaPlus, FaMinus, FaTag, FaSortNumericDown, FaAngleRight, FaCheck, FaTimes } from 'react-icons/fa';
+import React, { memo, useMemo, useRef, useCallback } from 'react';
+import { FaYinYang, FaSave, FaPlus, FaMinus, FaTag, FaSortNumericDown, FaAngleRight, FaRegCheckCircle, FaRegCircle } from 'react-icons/fa';
 import { AiOutlineSplitCells } from 'react-icons/ai';
 import { motion } from 'framer-motion';
-import Swal from 'sweetalert2';
 import { useTranslation, Trans } from 'react-i18next';
 import { ReactSortable } from 'react-sortablejs';
 import isEqual from 'lodash/isEqual';
 import useDebounce from 'react-use/lib/useDebounce';
 import scrollIntoView from 'scroll-into-view-if-needed';
 
+import Swal from './swal';
 import useContextMenu from './hooks/useContextMenu';
-import { saveColor } from './colors';
-import { getSegColors } from './util/colors';
+import useUserSettings from './hooks/useUserSettings';
+import { saveColor, controlsBackground, primaryTextColor, darkModeTransition } from './colors';
+import { useSegColors } from './contexts';
+import { mySpring } from './animations';
 
 const buttonBaseStyle = {
   margin: '0 3px', borderRadius: 3, color: 'white', cursor: 'pointer',
 };
 
-const neutralButtonColor = 'rgba(255, 255, 255, 0.2)';
+const neutralButtonColor = 'var(--gray8)';
 
 
-const Segment = memo(({ seg, index, currentSegIndex, formatTimecode, getFrameCount, updateOrder, invertCutSegments, onClick, onRemovePress, onReorderPress, onLabelPress, enabled, onExportSingleSegmentClick, onExportSegmentEnabledToggle, onExportSegmentDisableAll, onExportSegmentEnableAll, jumpSegStart, jumpSegEnd, addCutSegment, onViewSegmentTagsPress }) => {
+const Segment = memo(({ darkMode, seg, index, currentSegIndex, formatTimecode, getFrameCount, updateOrder, invertCutSegments, onClick, onRemovePress, onRemoveSelected, onLabelSelectedSegments, onReorderPress, onLabelPress, selected, onSelectSingleSegment, onToggleSegmentSelected, onDeselectAllSegments, onSelectSegmentsByLabel, onSelectAllSegments, jumpSegStart, jumpSegEnd, addSegment, onViewSegmentTags, onExtractSegmentFramesAsImages, onInvertSelectedSegments }) => {
   const { t } = useTranslation();
+  const { getSegColor } = useSegColors();
 
   const ref = useRef();
 
-  useContextMenu(ref, invertCutSegments ? [] : [
-    { label: t('Jump to cut start'), click: jumpSegStart },
-    { label: t('Jump to cut end'), click: jumpSegEnd },
+  const contextMenuTemplate = useMemo(() => {
+    if (invertCutSegments) return [];
+    return [
+      { label: t('Jump to start time'), click: jumpSegStart },
+      { label: t('Jump to end time'), click: jumpSegEnd },
 
-    { type: 'separator' },
+      { type: 'separator' },
 
-    { label: t('Add segment'), click: addCutSegment },
-    { label: t('Label segment'), click: onLabelPress },
-    { label: t('Remove segment'), click: onRemovePress },
+      { label: t('Add segment'), click: addSegment },
+      { label: t('Label segment'), click: onLabelPress },
+      { label: t('Remove segment'), click: onRemovePress },
 
-    { type: 'separator' },
+      { type: 'separator' },
 
-    { label: t('Change segment order'), click: onReorderPress },
-    { label: t('Increase segment order'), click: () => updateOrder(1) },
-    { label: t('Decrease segment order'), click: () => updateOrder(-1) },
+      { label: t('Select only this segment'), click: () => onSelectSingleSegment(seg) },
+      { label: t('Select all segments'), click: () => onSelectAllSegments() },
+      { label: t('Deselect all segments'), click: () => onDeselectAllSegments() },
+      { label: t('Select segments by label'), click: () => onSelectSegmentsByLabel(seg) },
+      { label: t('Invert selected segments'), click: () => onInvertSelectedSegments() },
 
-    { type: 'separator' },
+      { type: 'separator' },
 
-    { label: t('Include ONLY this segment in export'), click: () => onExportSingleSegmentClick(seg) },
-    { label: enabled ? t('Exclude this segment from export') : t('Include this segment in export'), click: () => onExportSegmentEnabledToggle(seg) },
-    { label: t('Include all segments in export'), click: () => onExportSegmentEnableAll(seg) },
-    { label: t('Exclude all segments from export'), click: () => onExportSegmentDisableAll(seg) },
+      { label: t('Label selected segments'), click: onLabelSelectedSegments },
+      { label: t('Remove selected segments'), click: onRemoveSelected },
 
-    { type: 'separator' },
+      { type: 'separator' },
 
-    { label: t('Segment tags'), click: () => onViewSegmentTagsPress(index) },
-  ]);
+      { label: t('Change segment order'), click: onReorderPress },
+      { label: t('Increase segment order'), click: () => updateOrder(1) },
+      { label: t('Decrease segment order'), click: () => updateOrder(-1) },
+
+      { type: 'separator' },
+
+      { label: t('Segment tags'), click: () => onViewSegmentTags(index) },
+      { label: t('Extract frames as image files'), click: () => onExtractSegmentFramesAsImages(index) },
+    ];
+  }, [invertCutSegments, t, jumpSegStart, jumpSegEnd, addSegment, onLabelPress, onRemovePress, onLabelSelectedSegments, onRemoveSelected, onReorderPress, onSelectSingleSegment, seg, onSelectAllSegments, onDeselectAllSegments, onSelectSegmentsByLabel, onInvertSelectedSegments, updateOrder, onViewSegmentTags, index, onExtractSegmentFramesAsImages]);
+
+  useContextMenu(ref, contextMenuTemplate);
 
   const duration = seg.end - seg.start;
   const durationMs = duration * 1000;
@@ -66,17 +80,20 @@ const Segment = memo(({ seg, index, currentSegIndex, formatTimecode, getFrameCou
   function renderNumber() {
     if (invertCutSegments) return <FaSave style={{ color: saveColor, marginRight: 5, verticalAlign: 'middle' }} size={14} />;
 
-    const { segBgColor, segBorderColor } = getSegColors(seg);
+    const segColor = getSegColor(seg);
 
-    return <b style={{ color: 'white', padding: '0 4px', marginRight: 3, background: segBgColor, border: `1px solid ${isActive ? segBorderColor : 'transparent'}`, borderRadius: 10, fontSize: 12 }}>{index + 1}</b>;
+    const color = segColor.desaturate(0.25).lightness(darkMode ? 35 : 55);
+    const borderColor = darkMode ? color.lighten(0.5) : color.darken(0.3);
+
+    return <b style={{ cursor: 'grab', color: 'white', padding: '0 4px', marginRight: 3, marginLeft: -3, background: color.string(), border: `1px solid ${isActive ? borderColor.string() : 'transparent'}`, borderRadius: 10, fontSize: 12 }}>{index + 1}</b>;
   }
 
-  const timeStr = useMemo(() => `${formatTimecode(seg.start)} - ${formatTimecode(seg.end)}`, [seg.start, seg.end, formatTimecode]);
+  const timeStr = useMemo(() => `${formatTimecode({ seconds: seg.start })} - ${formatTimecode({ seconds: seg.end })}`, [seg.start, seg.end, formatTimecode]);
 
   function onDoubleClick() {
     if (invertCutSegments) return;
-    if (!enabled) {
-      onExportSegmentEnabledToggle(seg);
+    if (!selected) {
+      onToggleSegmentSelected(seg);
       return;
     }
     jumpSegStart();
@@ -85,33 +102,44 @@ const Segment = memo(({ seg, index, currentSegIndex, formatTimecode, getFrameCou
   const durationMsFormatted = Math.floor(durationMs);
   const frameCount = getFrameCount(duration);
 
+  const CheckIcon = selected ? FaRegCheckCircle : FaRegCircle;
+
+  const onToggleSegmentSelectedClick = useCallback((e) => {
+    e.stopPropagation();
+    onToggleSegmentSelected(seg);
+  }, [onToggleSegmentSelected, seg]);
+
+  const cursor = invertCutSegments ? undefined : 'grab';
+
   return (
     <motion.div
       ref={ref}
       role="button"
       onClick={() => !invertCutSegments && onClick(index)}
       onDoubleClick={onDoubleClick}
-      positionTransition
-      style={{ cursor: 'grab', originY: 0, margin: '5px 0', background: 'rgba(0,0,0,0.1)', border: `1px solid rgba(255,255,255,${isActive ? 1 : 0.3})`, padding: 5, borderRadius: 5, position: 'relative', opacity: !enabled && !invertCutSegments ? 0.5 : undefined }}
+      layout
+      style={{ originY: 0, margin: '5px 0', background: 'var(--gray2)', border: isActive ? '1px solid var(--gray10)' : '1px solid transparent', padding: 5, borderRadius: 5, position: 'relative' }}
       initial={{ scaleY: 0 }}
-      animate={{ scaleY: 1 }}
+      animate={{ scaleY: 1, opacity: !selected && !invertCutSegments ? 0.5 : undefined }}
       exit={{ scaleY: 0 }}
+      className="segment-list-entry"
     >
-      <div style={{ color: 'white', marginBottom: 3, display: 'flex', alignItems: 'center', height: 16 }}>
+      <div className="segment-handle" style={{ cursor, color: 'var(--gray12)', marginBottom: 3, display: 'flex', alignItems: 'center', height: 16 }}>
         {renderNumber()}
-        <span style={{ fontSize: 310 / timeStr.length, whiteSpace: 'nowrap' }}>{timeStr}</span>
-      </div>
-      <div style={{ fontSize: 12, color: 'white' }}>{seg.name}</div>
-      <div style={{ fontSize: 13 }}>
-        {t('Duration')} {prettyMs(durationMs)}
-      </div>
-      <div style={{ fontSize: 12 }}>
-        <Trans>{{ durationMsFormatted }} ms, {{ frameCount }} frames</Trans>
+        <span style={{ cursor, fontSize: Math.min(310 / timeStr.length, 14), whiteSpace: 'nowrap' }}>{timeStr}</span>
       </div>
 
-      {!enabled && !invertCutSegments && (
-        <div style={{ position: 'absolute', pointerEvents: 'none', top: 0, right: 0, bottom: 0, left: 0, overflow: 'hidden', display: 'flex', justifyContent: 'center', alignItems: 'center' }}>
-          <FaTimes style={{ fontSize: 100, color: 'rgba(255,0,0,0.8)' }} />
+      <div style={{ fontSize: 12, color: primaryTextColor }}>{seg.name}</div>
+      <div style={{ fontSize: 13 }}>
+        {t('Duration')} {formatTimecode({ seconds: duration, shorten: true })}
+      </div>
+      <div style={{ fontSize: 12 }}>
+        <Trans>{{ durationMsFormatted }} ms, {{ frameCount: frameCount ?? '?' }} frames</Trans>
+      </div>
+
+      {!invertCutSegments && (
+        <div style={{ position: 'absolute', right: 3, bottom: 3 }}>
+          <CheckIcon className="enabled" size={20} color="var(--gray12)" onClick={onToggleSegmentSelectedClick} />
         </div>
       )}
     </motion.div>
@@ -119,39 +147,49 @@ const Segment = memo(({ seg, index, currentSegIndex, formatTimecode, getFrameCou
 });
 
 const SegmentList = memo(({
-  formatTimecode, cutSegments, outSegments, getFrameCount, onSegClick,
-  currentSegIndex, invertCutSegments,
-  updateSegOrder, updateSegOrders, addCutSegment, removeCutSegment,
-  onLabelSegmentPress, currentCutSeg, segmentAtCursor, toggleSideBar, splitCurrentSegment,
-  enabledOutSegments, enabledOutSegmentsRaw, onExportSingleSegmentClick, onExportSegmentEnabledToggle, onExportSegmentDisableAll, onExportSegmentEnableAll,
-  jumpSegStart, jumpSegEnd, simpleMode, onViewSegmentTagsPress,
+  width, formatTimecode, apparentCutSegments, inverseCutSegments, getFrameCount, onSegClick,
+  currentSegIndex,
+  updateSegOrder, updateSegOrders, addSegment, removeCutSegment, onRemoveSelected,
+  onLabelSegment, currentCutSeg, segmentAtCursor, toggleSegmentsList, splitCurrentSegment,
+  selectedSegments, isSegmentSelected, onSelectSingleSegment, onToggleSegmentSelected, onDeselectAllSegments, onSelectAllSegments, onSelectSegmentsByLabel, onExtractSegmentFramesAsImages, onLabelSelectedSegments, onInvertSelectedSegments,
+  jumpSegStart, jumpSegEnd, onViewSegmentTags,
 }) => {
   const { t } = useTranslation();
+  const { getSegColor } = useSegColors();
 
-  const sortableList = outSegments.map((seg) => ({ id: seg.segId, seg }));
+  const { invertCutSegments, simpleMode, darkMode } = useUserSettings();
 
-  function setSortableList(newList) {
-    if (isEqual(outSegments.map((s) => s.segId), newList.map((l) => l.id))) return; // No change
+  const segments = invertCutSegments ? inverseCutSegments : apparentCutSegments;
+
+  const sortableList = segments.map((seg) => ({ id: seg.segId, seg }));
+
+  const setSortableList = useCallback((newList) => {
+    if (isEqual(segments.map((s) => s.segId), newList.map((l) => l.id))) return; // No change
     updateSegOrders(newList.map((list) => list.id));
+  }, [segments, updateSegOrders]);
+
+  let header = t('Segments to export:');
+  if (segments.length === 0) {
+    if (invertCutSegments) {
+      header = (
+        <Trans>You have enabled the &quot;invert segments&quot; mode <FaYinYang style={{ verticalAlign: 'middle' }} /> which will cut away selected segments instead of keeping them. But there is no space between any segments, or at least two segments are overlapping. This would not produce any output. Either make room between segments or click the Yinyang <FaYinYang style={{ verticalAlign: 'middle', color: primaryTextColor }} /> symbol below to disable this mode. Alternatively you may combine overlapping segments from the menu.</Trans>
+      );
+    } else {
+      header = t('No segments to export.');
+    }
   }
 
-  let headerText = t('Segments to export:');
-  if (outSegments.length === 0) {
-    if (invertCutSegments) headerText = t('Make sure you have no overlapping segments.');
-    else headerText = t('No segments to export.');
-  }
-
-  async function onReorderSegsPress(index) {
-    if (cutSegments.length < 2) return;
+  async function onReorderSegs(index) {
+    if (apparentCutSegments.length < 2) return;
     const { value } = await Swal.fire({
       title: `${t('Change order of segment')} ${index + 1}`,
-      text: `Please enter a number from 1 to ${cutSegments.length} to be the new order for the current segment`,
+      text: `Please enter a number from 1 to ${apparentCutSegments.length} to be the new order for the current segment`,
       input: 'text',
       inputValue: index + 1,
       showCancelButton: true,
       inputValidator: (v) => {
         const parsed = parseInt(v, 10);
-        return Number.isNaN(parsed) || parsed > cutSegments.length || parsed < 1 ? t('Invalid number entered') : undefined;
+        return Number.isNaN(parsed) || parsed > apparentCutSegments.length || parsed < 1 ? t('Invalid number entered') : undefined;
       },
     });
 
@@ -162,30 +200,26 @@ const SegmentList = memo(({
   }
 
   function renderFooter() {
-    const { segActiveBgColor: currentSegActiveBgColor } = getSegColors(currentCutSeg);
-    const { segActiveBgColor: segmentAtCursorActiveBgColor } = getSegColors(segmentAtCursor);
+    const getButtonColor = (seg) => getSegColor(seg).desaturate(0.3).lightness(darkMode ? 45 : 55).string();
+    const currentSegColor = getButtonColor(currentCutSeg);
+    const segAtCursorColor = getButtonColor(segmentAtCursor);
 
-    function renderExportEnabledCheckBox() {
-      const segmentExportEnabled = currentCutSeg && enabledOutSegmentsRaw.some((s) => s.segId === currentCutSeg.segId);
-      const Icon = segmentExportEnabled ? FaCheck : FaTimes;
-
-      return <Icon size={24} title={segmentExportEnabled ? t('Include this segment in export') : t('Exclude this segment from export')} style={{ ...buttonBaseStyle, backgroundColor: currentSegActiveBgColor }} role="button" onClick={() => onExportSegmentEnabledToggle(currentCutSeg)} />;
-    }
+    const segmentsTotal = selectedSegments.reduce((acc, { start, end }) => (end - start) + acc, 0);
 
     return (
       <>
-        <div style={{ display: 'flex', padding: '5px 0', alignItems: 'center', justifyContent: 'center', borderBottom: '1px solid grey' }}>
+        <div style={{ display: 'flex', padding: '5px 0', alignItems: 'center', justifyContent: 'center', borderBottom: '1px solid var(gray6)' }}>
           <FaPlus
             size={24}
             style={{ ...buttonBaseStyle, background: neutralButtonColor }}
             role="button"
             title={t('Add segment')}
-            onClick={addCutSegment}
+            onClick={addSegment}
           />
 
           <FaMinus
             size={24}
-            style={{ ...buttonBaseStyle, background: cutSegments.length >= 2 ? currentSegActiveBgColor : neutralButtonColor }}
+            style={{ ...buttonBaseStyle, background: apparentCutSegments.length >= 2 ? currentSegColor : neutralButtonColor }}
             role="button"
             title={`${t('Remove segment')} ${currentSegIndex + 1}`}
             onClick={() => removeCutSegment(currentSegIndex)}
@@ -197,68 +231,74 @@ const SegmentList = memo(({
                 size={16}
                 title={t('Change segment order')}
                 role="button"
-                style={{ ...buttonBaseStyle, padding: 4, background: currentSegActiveBgColor }}
-                onClick={() => onReorderSegsPress(currentSegIndex)}
+                style={{ ...buttonBaseStyle, padding: 4, background: currentSegColor }}
+                onClick={() => onReorderSegs(currentSegIndex)}
               />
 
               <FaTag
                 size={16}
                 title={t('Label segment')}
                 role="button"
-                style={{ ...buttonBaseStyle, padding: 4, background: currentSegActiveBgColor }}
-                onClick={() => onLabelSegmentPress(currentSegIndex)}
+                style={{ ...buttonBaseStyle, padding: 4, background: currentSegColor }}
+                onClick={() => onLabelSegment(currentSegIndex)}
               />
-
-              {renderExportEnabledCheckBox()}
             </>
           )}
 
           <AiOutlineSplitCells
-            size={16}
+            size={22}
             title={t('Split segment at cursor')}
             role="button"
-            style={{ ...buttonBaseStyle, padding: 4, background: segmentAtCursor ? segmentAtCursorActiveBgColor : neutralButtonColor }}
+            style={{ ...buttonBaseStyle, padding: 1, background: segmentAtCursor ? segAtCursorColor : neutralButtonColor }}
             onClick={splitCurrentSegment}
           />
         </div>
 
-        <div style={{ padding: '5px 10px', boxSizing: 'border-box', borderBottom: '1px solid grey', borderTop: '1px solid grey', display: 'flex', justifyContent: 'space-between', fontSize: 13 }}>
+        <div style={{ padding: '5px 10px', boxSizing: 'border-box', borderBottom: '1px solid var(gray6)', borderTop: '1px solid var(gray6)', display: 'flex', justifyContent: 'space-between', fontSize: 13 }}>
           <div>{t('Segments total:')}</div>
-          <div>{formatTimecode(enabledOutSegments.reduce((acc, { start, end }) => (end - start) + acc, 0))}</div>
+          <div>{formatTimecode({ seconds: segmentsTotal })}</div>
         </div>
       </>
     );
   }
 
   return (
-    <>
+    <motion.div
+      style={{ width, background: controlsBackground, borderLeft: '1px solid var(--gray7)', color: 'var(--gray11)', transition: darkModeTransition, display: 'flex', flexDirection: 'column', overflowY: 'hidden' }}
+      initial={{ x: width }}
+      animate={{ x: 0 }}
+      exit={{ x: width }}
+      transition={mySpring}
+    >
+      <div style={{ fontSize: 14, padding: '0 5px', color: 'var(--gray12)' }} className="no-user-select">
+        <FaAngleRight
+          title={t('Close sidebar')}
+          size={20}
+          style={{ verticalAlign: 'middle', color: 'var(--gray11)', cursor: 'pointer', padding: 2 }}
+          role="button"
+          onClick={toggleSegmentsList}
+        />
+
+        {header}
+      </div>
+
       <div style={{ padding: '0 10px', overflowY: 'scroll', flexGrow: 1 }} className="hide-scrollbar">
-        <div style={{ fontSize: 14, marginBottom: 10 }}>
-          <FaAngleRight
-            title={t('Close sidebar')}
-            size={18}
-            style={{ verticalAlign: 'middle', color: 'white', cursor: 'pointer' }}
-            role="button"
-            onClick={toggleSideBar}
-          />
-
-          {headerText}
-        </div>
-
-        <ReactSortable list={sortableList} setList={setSortableList} sort={!invertCutSegments}>
+        <ReactSortable list={sortableList} setList={setSortableList} disabled={!!invertCutSegments} handle=".segment-handle">
           {sortableList.map(({ id, seg }, index) => {
-            const enabled = !invertCutSegments && enabledOutSegmentsRaw.includes(seg);
+            const selected = !invertCutSegments && isSegmentSelected({ segId: seg.segId });
             return (
               <Segment
                 key={id}
+                darkMode={darkMode}
                 seg={seg}
                 index={index}
-                enabled={enabled}
+                selected={selected}
                 onClick={onSegClick}
-                addCutSegment={addCutSegment}
+                addSegment={addSegment}
+                onRemoveSelected={onRemoveSelected}
                 onRemovePress={() => removeCutSegment(index)}
-                onReorderPress={() => onReorderSegsPress(index)}
-                onLabelPress={() => onLabelSegmentPress(index)}
+                onReorderPress={() => onReorderSegs(index)}
+                onLabelPress={() => onLabelSegment(index)}
                 jumpSegStart={() => jumpSegStart(index)}
                 jumpSegEnd={() => jumpSegEnd(index)}
                 updateOrder={(dir) => updateSegOrder(index, index + dir)}
@@ -266,19 +306,23 @@ const SegmentList = memo(({
                 formatTimecode={formatTimecode}
                 currentSegIndex={currentSegIndex}
                 invertCutSegments={invertCutSegments}
-                onExportSingleSegmentClick={onExportSingleSegmentClick}
-                onExportSegmentEnabledToggle={onExportSegmentEnabledToggle}
-                onExportSegmentDisableAll={onExportSegmentDisableAll}
-                onExportSegmentEnableAll={onExportSegmentEnableAll}
-                onViewSegmentTagsPress={onViewSegmentTagsPress}
+                onSelectSingleSegment={onSelectSingleSegment}
+                onToggleSegmentSelected={onToggleSegmentSelected}
+                onDeselectAllSegments={onDeselectAllSegments}
+                onSelectAllSegments={onSelectAllSegments}
+                onViewSegmentTags={onViewSegmentTags}
+                onSelectSegmentsByLabel={onSelectSegmentsByLabel}
+                onExtractSegmentFramesAsImages={onExtractSegmentFramesAsImages}
+                onLabelSelectedSegments={onLabelSelectedSegments}
+                onInvertSelectedSegments={onInvertSelectedSegments}
               />
             );
           })}
         </ReactSortable>
       </div>
 
-      {outSegments.length > 0 && renderFooter()}
-    </>
+      {segments.length > 0 && renderFooter()}
+    </motion.div>
   );
 });
 
