@@ -1,21 +1,37 @@
-import React, { memo } from 'react';
+import React, { memo, useMemo } from 'react';
 import { motion, AnimatePresence } from 'framer-motion';
 import { FaTrashAlt } from 'react-icons/fa';
 
 import { mySpring } from './animations';
-
-import { formatDuration } from './util/duration';
+import useUserSettings from './hooks/useUserSettings';
+import { useSegColors } from './contexts';
 
 
 const TimelineSeg = memo(({
-  duration, cutStart, cutEnd, isActive, segNum, name,
-  onSegClick, invertCutSegments, segBgColor, segActiveBgColor, segBorderColor,
+  seg, duration, isActive, segNum, onSegClick, invertCutSegments, formatTimecode, selected,
 }) => {
+  const { darkMode } = useUserSettings();
+  const { getSegColor } = useSegColors();
+
+  const segColor = useMemo(() => getSegColor(seg), [getSegColor, seg]);
+
+  const { name, start: cutStart, end: cutEnd } = seg;
+
   const cutSectionWidth = `${((cutEnd - cutStart) / duration) * 100}%`;
 
   const startTimePos = `${(cutStart / duration) * 100}%`;
-  const markerBorder = `2px solid ${isActive ? segBorderColor : 'transparent'}`;
-  const backgroundColor = isActive ? segActiveBgColor : segBgColor;
+
+  const markerBorder = useMemo(() => {
+    if (!isActive) return '2px solid transparent';
+    return `1.5px solid ${darkMode ? segColor.desaturate(0.1).lightness(70).string() : segColor.desaturate(0.2).lightness(40).string()}`;
+  }, [darkMode, isActive, segColor]);
+
+  const backgroundColor = useMemo(() => {
+    // we use both transparency and lightness, so that segments can be visible when overlapping
+    if (invertCutSegments || !selected) return darkMode ? segColor.desaturate(0.3).lightness(30).alpha(0.5).string() : segColor.desaturate(0.3).lightness(70).alpha(0.5).string();
+    if (isActive) return darkMode ? segColor.saturate(0.2).lightness(60).alpha(0.7).string() : segColor.saturate(0.2).lightness(40).alpha(0.8).string();
+    return darkMode ? segColor.desaturate(0.2).lightness(50).alpha(0.7).string() : segColor.lightness(35).alpha(0.6).string();
+  }, [darkMode, invertCutSegments, isActive, segColor, selected]);
   const markerBorderRadius = 5;
 
   const wrapperStyle = {
@@ -27,9 +43,9 @@ const TimelineSeg = memo(({
     display: 'flex',
     alignItems: 'center',
     justifyContent: 'space-between',
-    background: backgroundColor,
     originX: 0,
     boxSizing: 'border-box',
+    color: 'white',
 
     borderLeft: markerBorder,
     borderTopLeftRadius: markerBorderRadius,
@@ -42,19 +58,21 @@ const TimelineSeg = memo(({
 
   const onThisSegClick = () => onSegClick(segNum);
 
-  const durationStr = cutEnd > cutStart ? `${formatDuration({ seconds: cutEnd - cutStart, shorten: true })} ` : '';
-  const title = `${durationStr}${name}`;
+  const title = [];
+  if (cutEnd > cutStart) title.push(`${formatTimecode({ seconds: cutEnd - cutStart, shorten: true })}`);
+  if (name) title.push(name);
 
   return (
     <motion.div
       style={wrapperStyle}
-      layoutTransition={mySpring}
+      layout
+      transition={mySpring}
       initial={{ opacity: 0, scaleX: 0 }}
-      animate={{ opacity: 1, scaleX: 1 }}
+      animate={{ opacity: 1, scaleX: 1, backgroundColor }}
       exit={{ opacity: 0, scaleX: 0 }}
       role="button"
       onClick={onThisSegClick}
-      title={title}
+      title={title.join(' ')}
     >
       <div style={{ alignSelf: 'flex-start', flexShrink: 1, fontSize: 10, minWidth: 0, overflow: 'hidden' }}>{segNum + 1}</div>
 
@@ -67,7 +85,7 @@ const TimelineSeg = memo(({
             style={{ width: 16, height: 16, flexShrink: 1 }}
           >
             <FaTrashAlt
-              style={{ width: '100%', color: segBorderColor }}
+              style={{ width: '100%', color: 'var(--gray12)' }}
               size={16}
             />
           </motion.div>
