@@ -1,24 +1,27 @@
-import React, { memo } from 'react';
-import { IoIosHelpCircle, IoIosSettings } from 'react-icons/io';
+import React, { memo, useCallback } from 'react';
+import { IoIosSettings } from 'react-icons/io';
 import { FaLock, FaUnlock } from 'react-icons/fa';
 import { IconButton, Button, CrossIcon, ListIcon, VolumeUpIcon, VolumeOffIcon } from 'evergreen-ui';
 import { useTranslation } from 'react-i18next';
 
-import MergeExportButton from './components/MergeExportButton';
+import ExportModeButton from './components/ExportModeButton';
 
-import { withBlur, isMasBuild } from './util';
-import { primaryTextColor } from './colors';
+import { withBlur } from './util';
+import { primaryTextColor, controlsBackground, darkModeTransition } from './colors';
+import useUserSettings from './hooks/useUserSettings';
 
 
 const TopMenu = memo(({
-  filePath, copyAnyAudioTrack, toggleStripAudio, customOutDir, changeOutDir,
-  renderOutFmt, toggleHelp, numStreamsToCopy, numStreamsTotal, setStreamsSelectorShown, toggleSettings,
-  enabledOutSegments, autoMerge, setAutoMerge, autoDeleteMergedSegments, setAutoDeleteMergedSegments, isCustomFormatSelected, onOutFormatLockedClick, simpleMode, outFormatLocked, clearOutDir,
+  filePath, fileFormat, copyAnyAudioTrack, toggleStripAudio,
+  renderOutFmt, numStreamsToCopy, numStreamsTotal, setStreamsSelectorShown, toggleSettings,
+  selectedSegments, isCustomFormatSelected, clearOutDir,
 }) => {
   const { t } = useTranslation();
+  const { customOutDir, changeOutDir, simpleMode, outFormatLocked, setOutFormatLocked } = useUserSettings();
 
-  // We cannot allow exporting to a directory which has not yet been confirmed by an open dialog because of sandox restrictions
-  const showClearWorkingDirButton = customOutDir && !isMasBuild;
+  const onOutFormatLockedClick = useCallback(() => setOutFormatLocked((v) => (v ? undefined : fileFormat)), [fileFormat, setOutFormatLocked]);
+
+  const showClearWorkingDirButton = !!customOutDir;
 
   function renderFormatLock() {
     const Icon = outFormatLocked ? FaLock : FaUnlock;
@@ -26,7 +29,10 @@ const TopMenu = memo(({
   }
 
   return (
-    <>
+    <div
+      className="no-user-select"
+      style={{ background: controlsBackground, transition: darkModeTransition, display: 'flex', alignItems: 'center', padding: '3px 5px', justifyContent: 'space-between', flexWrap: 'wrap' }}
+    >
       {filePath && (
         <>
           <Button height={20} iconBefore={ListIcon} onClick={withBlur(() => setStreamsSelectorShown(true))}>
@@ -36,7 +42,7 @@ const TopMenu = memo(({
           <Button
             iconBefore={copyAnyAudioTrack ? VolumeUpIcon : VolumeOffIcon}
             height={20}
-            title={`${t('Discard audio? Current:')} ${copyAnyAudioTrack ? t('Keep audio tracks') : t('Discard audio tracks')}`}
+            title={copyAnyAudioTrack ? t('Keep audio tracks') : t('Discard audio tracks')}
             onClick={withBlur(toggleStripAudio)}
           >
             {copyAnyAudioTrack ? t('Keep audio') : t('Discard audio')}
@@ -71,13 +77,12 @@ const TopMenu = memo(({
 
           {!simpleMode && (isCustomFormatSelected || outFormatLocked) && renderFormatLock()}
 
-          <MergeExportButton autoMerge={autoMerge} enabledOutSegments={enabledOutSegments} setAutoMerge={setAutoMerge} autoDeleteMergedSegments={autoDeleteMergedSegments} setAutoDeleteMergedSegments={setAutoDeleteMergedSegments} />
+          <ExportModeButton selectedSegments={selectedSegments} style={{ flexGrow: 0, flexBasis: 140 }} />
         </>
       )}
 
-      <IoIosHelpCircle size={24} role="button" onClick={toggleHelp} style={{ verticalAlign: 'middle', marginLeft: 5 }} />
       <IoIosSettings size={24} role="button" onClick={toggleSettings} style={{ verticalAlign: 'middle', marginLeft: 5 }} />
-    </>
+    </div>
   );
 });
 
