@@ -1,2 +1,2 @@
 // eslint-disable-next-line import/prefer-default-export
-export const mySpring = { type: 'spring', damping: 25, stiffness: 350 };
+export const mySpring = { type: 'spring', damping: 50, stiffness: 700 };
