@@ -1,6 +1,8 @@
-export const saveColor = 'hsl(158, 100%, 43%)';
-export const primaryColor = 'hsl(194, 78%, 47%)';
-export const primaryTextColor = 'hsla(194, 100%, 66%, 1)';
-export const waveformColor = '#ffffff'; // Must be hex because used by ffmpeg
-export const controlsBackground = '#6b6b6b';
-export const timelineBackground = '#444';
+export const saveColor = 'var(--green11)';
+export const primaryColor = 'var(--cyan9)';
+export const primaryTextColor = 'var(--cyan11)';
+export const waveformColorLight = '#000000'; // Must be hex because used by ffmpeg
+export const waveformColorDark = '#ffffff'; // Must be hex because used by ffmpeg
+export const controlsBackground = 'var(--gray4)';
+export const timelineBackground = 'var(--gray2)';
+export const darkModeTransition = 'background .5s';
