@@ -1,6 +1,10 @@
 import React, { memo } from 'react';
 
+import { primaryTextColor } from '../colors';
+
+export const highlightedTextStyle = { textDecoration: 'underline', textUnderlineOffset: '.2em', textDecorationColor: primaryTextColor, color: 'var(--gray12)', borderRadius: '.4em' };
+
 // eslint-disable-next-line react/jsx-props-no-spreading
-const HighlightedText = memo(({ children, style, ...props }) => <span {...props} style={{ background: 'rgb(193, 98, 0)', borderRadius: '.4em', padding: '0 .3em', ...style }}>{children}</span>);
+const HighlightedText = memo(({ children, style, ...props }) => <span {...props} style={{ ...highlightedTextStyle, ...style }}>{children}</span>);
 
 export default HighlightedText;
