@@ -1,17 +1,15 @@
 import React, { memo } from 'react';
-import { Button } from 'evergreen-ui';
-import { useTranslation } from 'react-i18next';
 
 import { withBlur } from '../util';
+import useUserSettings from '../hooks/useUserSettings';
+import Switch from './Switch';
 
 
-const PreserveMovDataButton = memo(({ preserveMovData, togglePreserveMovData }) => {
-  const { t } = useTranslation();
+const PreserveMovDataButton = memo(() => {
+  const { preserveMovData, togglePreserveMovData } = useUserSettings();
 
   return (
-    <Button height={20} onClick={withBlur(togglePreserveMovData)}>
-      {preserveMovData ? t('Yes') : t('No')}
-    </Button>
+    <Switch checked={preserveMovData} onCheckedChange={withBlur(togglePreserveMovData)} />
   );
 });
 
