@@ -1,19 +1,34 @@
 import pMap from 'p-map';
-import flatMap from 'lodash/flatMap';
 import sortBy from 'lodash/sortBy';
 import moment from 'moment';
 import i18n from 'i18next';
 import Timecode from 'smpte-timecode';
+import minBy from 'lodash/minBy';
 
-import { getOutPath, isDurationValid, getExtensionForFormat, isMac, isWindows } from './util';
+import { pcmAudioCodecs, getMapStreamsArgs, isMov } from './util/streams';
+import { getSuffixedOutPath, isWindows, isMac, platform, arch, isExecaFailure } from './util';
+import { isDurationValid } from './segments';
+
+import isDev from './isDev';
 
 const execa = window.require('execa');
 const { join } = window.require('path');
-const fileType = window.require('file-type');
-const readChunk = window.require('read-chunk');
+const FileType = window.require('file-type');
 const readline = window.require('readline');
-const isDev = window.require('electron-is-dev');
+const { pathExists } = window.require('fs-extra');
 
+let customFfPath;
+
+const runningFfmpegs = new Set();
+// setInterval(() => console.log(runningFfmpegs.size), 1000);
+
+
+export class RefuseOverwriteError extends Error {}
+
+// Note that this does not work on MAS because of sandbox restrictions
+export function setCustomFfPath(path) {
+  customFfPath = path;
+}
 
 export function getFfCommandLine(cmd, args) {
   const mapArg = arg => (/[^0-9a-zA-Z-_]/.test(arg) ? `'${arg}'` : arg);
@@ -21,36 +36,69 @@ export function getFfCommandLine(cmd, args) {
 }
 
 function getFfPath(cmd) {
-  // Testing non-mac setup on mac:
-  // return `node_modules/ffmpeg-ffprobe-static/${cmd}`;
-
-  if (isMac) {
-    return isDev ? `ffmpeg-mac/${cmd}` : join(window.process.resourcesPath, cmd);
-  }
-
   const exeName = isWindows ? `${cmd}.exe` : cmd;
-  return isDev
-    ? `node_modules/ffmpeg-ffprobe-static/${exeName}`
-    : join(window.process.resourcesPath, `node_modules/ffmpeg-ffprobe-static/${exeName}`);
+
+  if (customFfPath) return join(customFfPath, exeName);
+  if (isDev) return join('ffmpeg', `${platform}-${arch}`, exeName);
+  return join(window.process.resourcesPath, exeName);
 }
 
 export const getFfmpegPath = () => getFfPath('ffmpeg');
 export const getFfprobePath = () => getFfPath('ffprobe');
 
-export async function runFfprobe(args) {
+export async function runFfprobe(args, { timeout = isDev ? 10000 : 30000 } = {}) {
   const ffprobePath = getFfprobePath();
   console.log(getFfCommandLine('ffprobe', args));
-  return execa(ffprobePath, args);
+  const ps = execa(ffprobePath, args);
+  const timer = setTimeout(() => {
+    console.warn('killing timed out ffprobe');
+    ps.kill();
+  }, timeout);
+  try {
+    return await ps;
+  } finally {
+    clearTimeout(timer);
+  }
 }
 
-export function runFfmpeg(args) {
+// todo collect warnings from ffmpeg output and show them after export? example: https://github.com/mifi/lossless-cut/issues/1469
+export function runFfmpeg(args, execaOptions, { logCli = true } = {}) {
   const ffmpegPath = getFfmpegPath();
-  console.log(getFfCommandLine('ffmpeg', args));
-  return execa(ffmpegPath, args);
+  if (logCli) console.log(getFfCommandLine('ffmpeg', args));
+  const process = execa(ffmpegPath, args, execaOptions);
+
+  (async () => {
+    runningFfmpegs.add(process);
+    try {
+      await process;
+    } catch (err) {
+      // ignored here
+    } finally {
+      runningFfmpegs.delete(process);
+    }
+  })();
+  return process;
 }
 
+export function logStdoutStderr({ stdout, stderr }) {
+  if (stdout.length > 0) {
+    console.log('%cSTDOUT:', 'color: green; font-weight: bold');
+    console.log(stdout);
+  }
+  if (stderr.length > 0) {
+    console.log('%cSTDERR:', 'color: blue; font-weight: bold');
+    console.log(stderr);
+  }
+}
 
-export function handleProgress(process, cutDuration, onProgress) {
+export function abortFfmpegs() {
+  runningFfmpegs.forEach((process) => {
+    process.kill('SIGTERM', { forceKillAfterTimeout: 10000 });
+  });
+}
+
+export function handleProgress(process, durationIn, onProgress, customMatcher = () => {}) {
+  if (!onProgress) return;
   onProgress(0);
 
   const rl = readline.createInterface({ input: process.stderr });
@@ -61,13 +109,20 @@ export function handleProgress(process, cutDuration, onProgress) {
       let match = line.match(/frame=\s*[^\s]+\s+fps=\s*[^\s]+\s+q=\s*[^\s]+\s+(?:size|Lsize)=\s*[^\s]+\s+time=\s*([^\s]+)\s+/);
       // Audio only looks like this: "line size=  233422kB time=01:45:50.68 bitrate= 301.1kbits/s speed= 353x    "
       if (!match) match = line.match(/(?:size|Lsize)=\s*[^\s]+\s+time=\s*([^\s]+)\s+/);
-      if (!match) return;
+      if (!match) {
+        customMatcher(line);
+        return;
+      }
 
       const str = match[1];
       // console.log(str);
       const progressTime = Math.max(0, moment.duration(str).asSeconds());
       // console.log(progressTime);
-      const progress = cutDuration ? progressTime / cutDuration : 0;
+
+      if (durationIn == null) return;
+      const duration = Math.max(0, durationIn);
+      if (duration === 0) return;
+      const progress = duration ? Math.min(progressTime / duration, 1) : 0; // sometimes progressTime will be greater than cutDuration
       onProgress(progress);
     } catch (err) {
       console.log('Failed to parse ffmpeg progress line', err);
@@ -91,20 +146,59 @@ function getIntervalAroundTime(time, window) {
   };
 }
 
-export async function readFrames({ filePath, aroundTime, window, stream }) {
-  let intervalsArgs = [];
-  if (aroundTime != null) {
-    const { from, to } = getIntervalAroundTime(aroundTime, window);
-    intervalsArgs = ['-read_intervals', `${from}%${to}`];
-  }
-  const { stdout } = await runFfprobe(['-v', 'error', ...intervalsArgs, '-show_packets', '-select_streams', stream, '-show_entries', 'packet=pts_time,flags', '-of', 'json', filePath]);
+export async function readFrames({ filePath, from, to, streamIndex }) {
+  const intervalsArgs = from != null && to != null ? ['-read_intervals', `${from}%${to}`] : [];
+  const { stdout } = await runFfprobe(['-v', 'error', ...intervalsArgs, '-show_packets', '-select_streams', streamIndex, '-show_entries', 'packet=pts_time,flags', '-of', 'json', filePath]);
   const packetsFiltered = JSON.parse(stdout).packets
-    .map(p => ({ keyframe: p.flags[0] === 'K', time: parseFloat(p.pts_time, 10) }))
+    .map(p => ({
+      keyframe: p.flags[0] === 'K',
+      time: parseFloat(p.pts_time, 10),
+      createdAt: new Date(),
+    }))
     .filter(p => !Number.isNaN(p.time));
 
   return sortBy(packetsFiltered, 'time');
 }
 
+export async function readFramesAroundTime({ filePath, streamIndex, aroundTime, window }) {
+  if (aroundTime == null) throw new Error('aroundTime was nullish');
+  const { from, to } = getIntervalAroundTime(aroundTime, window);
+  return readFrames({ filePath, from, to, streamIndex });
+}
+
+export async function readKeyframesAroundTime({ filePath, streamIndex, aroundTime, window }) {
+  const frames = await readFramesAroundTime({ filePath, aroundTime, streamIndex, window });
+  return frames.filter((frame) => frame.keyframe);
+}
+
+export const findKeyframeAtExactTime = (keyframes, time) => keyframes.find((keyframe) => Math.abs(keyframe.time - time) < 0.000001);
+export const findNextKeyframe = (keyframes, time) => keyframes.find((keyframe) => keyframe.time >= time); // (assume they are already sorted)
+const findPreviousKeyframe = (keyframes, time) => keyframes.findLast((keyframe) => keyframe.time <= time);
+const findNearestKeyframe = (keyframes, time) => minBy(keyframes, (keyframe) => Math.abs(keyframe.time - time));
+
+function findKeyframe(keyframes, time, mode) {
+  switch (mode) {
+    case 'nearest': return findNearestKeyframe(keyframes, time);
+    case 'before': return findPreviousKeyframe(keyframes, time);
+    case 'after': return findNextKeyframe(keyframes, time);
+    default: return undefined;
+  }
+}
+
+export async function findKeyframeNearTime({ filePath, streamIndex, time, mode }) {
+  let keyframes = await readKeyframesAroundTime({ filePath, streamIndex, aroundTime: time, window: 10 });
+  let nearByKeyframe = findKeyframe(keyframes, time, mode);
+
+  if (!nearByKeyframe) {
+    keyframes = await readKeyframesAroundTime({ filePath, streamIndex, aroundTime: time, window: 60 });
+    nearByKeyframe = findKeyframe(keyframes, time, mode);
+  }
+
+  if (!nearByKeyframe) return undefined;
+  return nearByKeyframe.time;
+}
+
+// todo this is not in use
 // https://stackoverflow.com/questions/14005110/how-to-split-a-video-using-ffmpeg-so-that-each-chunk-starts-with-a-key-frame
 // http://kicherer.org/joomla/index.php/de/blog/42-avcut-frame-accurate-video-cutting-with-only-small-quality-loss
 export function getSafeCutTime(frames, cutTime, nextMode) {
@@ -158,15 +252,14 @@ export function findNearestKeyFrameTime({ frames, time, direction, fps }) {
   const sigma = fps ? (1 / fps) : 0.1;
   const keyframes = frames.filter(f => f.keyframe && (direction > 0 ? f.time > time + sigma : f.time < time - sigma));
   if (keyframes.length === 0) return undefined;
-  const nearestFrame = sortBy(keyframes, keyframe => (direction > 0 ? keyframe.time - time : time - keyframe.time))[0];
-  if (!nearestFrame) return undefined;
-  return nearestFrame.time;
+  const nearestKeyFrame = sortBy(keyframes, keyframe => (direction > 0 ? keyframe.time - time : time - keyframe.time))[0];
+  if (!nearestKeyFrame) return undefined;
+  return nearestKeyFrame.time;
 }
 
-export async function tryReadChaptersToEdl(filePath) {
+export async function tryMapChaptersToEdl(chapters) {
   try {
-    const { stdout } = await runFfprobe(['-i', filePath, '-show_chapters', '-print_format', 'json', '-hide_banner']);
-    return JSON.parse(stdout).chapters.map((chapter) => {
+    return chapters.map((chapter) => {
       const start = parseFloat(chapter.start_time);
       const end = parseFloat(chapter.end_time);
       if (Number.isNaN(start) || Number.isNaN(end)) return undefined;
@@ -185,8 +278,8 @@ export async function tryReadChaptersToEdl(filePath) {
   }
 }
 
-export async function getFormatData(filePath) {
-  console.log('getFormatData', filePath);
+async function readFormatData(filePath) {
+  console.log('readFormatData', filePath);
 
   const { stdout } = await runFfprobe([
     '-of', 'json', '-show_format', '-i', filePath, '-hide_banner',
@@ -196,87 +289,128 @@ export async function getFormatData(filePath) {
 
 
 export async function getDuration(filePath) {
-  return parseFloat((await getFormatData(filePath)).duration);
+  return parseFloat((await readFormatData(filePath)).duration);
 }
 
 export async function createChaptersFromSegments({ segmentPaths, chapterNames }) {
-  if (chapterNames) {
-    try {
-      const durations = await pMap(segmentPaths, (segmentPath) => getDuration(segmentPath), { concurrency: 3 });
-      let timeAt = 0;
-      return durations.map((duration, i) => {
-        const ret = { start: timeAt, end: timeAt + duration, name: chapterNames[i] };
-        timeAt += duration;
-        return ret;
-      });
-    } catch (err) {
-      console.error('Failed to create chapters from segments', err);
-    }
+  if (!chapterNames) return undefined;
+  try {
+    const durations = await pMap(segmentPaths, (segmentPath) => getDuration(segmentPath), { concurrency: 3 });
+    let timeAt = 0;
+    return durations.map((duration, i) => {
+      const ret = { start: timeAt, end: timeAt + duration, name: chapterNames[i] };
+      timeAt += duration;
+      return ret;
+    });
+  } catch (err) {
+    console.error('Failed to create chapters from segments', err);
+    return undefined;
   }
-  return undefined;
 }
 
 /**
  * ffmpeg only supports encoding certain formats, and some of the detected input
- * formats are not the same as the names used for encoding.
- * Therefore we have to map between detected format and encode format
+ * formats are not the same as the muxer name used for encoding.
+ * Therefore we have to map between detected input format and encode format
  * See also ffmpeg -formats
  */
-function mapFormat(requestedFormat) {
-  switch (requestedFormat) {
+function mapDefaultFormat({ streams, requestedFormat }) {
+  if (requestedFormat === 'mp4') {
+    // Only MOV supports these codecs, so default to MOV instead https://github.com/mifi/lossless-cut/issues/948
+    if (streams.some((stream) => pcmAudioCodecs.includes(stream.codec_name))) {
+      return 'mov';
+    }
+  }
+
+  // see sample.aac
+  if (requestedFormat === 'aac') return 'adts';
+
+  return requestedFormat;
+}
+
+async function determineOutputFormat(ffprobeFormatsStr, filePath) {
+  const ffprobeFormats = (ffprobeFormatsStr || '').split(',').map((str) => str.trim()).filter((str) => str);
+  if (ffprobeFormats.length === 0) {
+    console.warn('ffprobe returned unknown formats', ffprobeFormatsStr);
+    return undefined;
+  }
+
+  const [firstFfprobeFormat] = ffprobeFormats;
+  if (ffprobeFormats.length === 1) return firstFfprobeFormat;
+
+  // If ffprobe returned a list of formats, try to be a bit smarter about it.
+  // This should only be the case for matroska and mov. See `ffmpeg -formats`
+  if (!['matroska', 'mov'].includes(firstFfprobeFormat)) {
+    console.warn('Unknown ffprobe format list', ffprobeFormats);
+    return firstFfprobeFormat;
+  }
+
+  const fileTypeResponse = await FileType.fromFile(filePath);
+  if (fileTypeResponse == null) {
+    console.warn('file-type failed to detect format, defaulting to first', ffprobeFormats);
+    return firstFfprobeFormat;
+  }
+
+  // https://github.com/sindresorhus/file-type/blob/main/core.js
+  // https://www.ftyps.com/
+  // https://exiftool.org/TagNames/QuickTime.html
+  switch (fileTypeResponse.mime) {
+    case 'video/x-matroska': return 'matroska';
+    case 'video/webm': return 'webm';
+    case 'video/quicktime': return 'mov';
+    case 'video/3gpp2': return '3g2';
+    case 'video/3gpp': return '3gp';
+
     // These two cmds produce identical output, so we assume that encoding "ipod" means encoding m4a
     // ffmpeg -i example.aac -c copy OutputFile2.m4a
     // ffmpeg -i example.aac -c copy -f ipod OutputFile.m4a
     // See also https://github.com/mifi/lossless-cut/issues/28
-    case 'm4a': return 'ipod';
-    case 'aac': return 'ipod';
-    default: return requestedFormat;
+    case 'audio/x-m4a':
+    case 'audio/mp4':
+      return 'ipod';
+    case 'image/avif':
+    case 'image/heif':
+    case 'image/heif-sequence':
+    case 'image/heic':
+    case 'image/heic-sequence':
+    case 'video/x-m4v':
+    case 'video/mp4':
+    case 'image/x-canon-cr3':
+      return 'mp4';
+
+    default: {
+      console.warn('file-type returned unknown format', ffprobeFormats, fileTypeResponse.mime);
+      return firstFfprobeFormat;
+    }
   }
 }
 
-function determineOutputFormat(ffprobeFormats, ft) {
-  if (ffprobeFormats.includes(ft.ext)) return ft.ext;
-  return ffprobeFormats[0] || undefined;
-}
+export async function getSmarterOutFormat({ filePath, fileMeta: { format, streams } }) {
+  const formatsStr = format.format_name;
+  const assumedFormat = await determineOutputFormat(formatsStr, filePath);
 
-export async function getDefaultOutFormat(filePath, formatData) {
-  const formatsStr = formatData.format_name;
-  console.log('formats', formatsStr);
-  const formats = (formatsStr || '').split(',');
-
-  // ffprobe sometimes returns a list of formats, try to be a bit smarter about it.
-  const bytes = await readChunk(filePath, 0, 4100);
-  const ft = fileType(bytes) || {};
-  console.log(`fileType detected format ${JSON.stringify(ft)}`);
-  let assumedFormat = determineOutputFormat(formats, ft);
-
-  // https://github.com/mifi/lossless-cut/issues/367
-  if (assumedFormat === 'mp4' && formatData.tags && formatData.tags.major_brand === 'XAVC') assumedFormat = 'mov';
-
-  return mapFormat(assumedFormat);
-}
-
-export async function getAllStreams(filePath) {
-  const { stdout } = await runFfprobe([
-    '-of', 'json', '-show_entries', 'stream', '-i', filePath, '-hide_banner',
-  ]);
-
-  return JSON.parse(stdout);
+  return mapDefaultFormat({ streams, requestedFormat: assumedFormat });
 }
 
 export async function readFileMeta(filePath) {
   try {
-    const [formatData, allStreamsResponse] = await Promise.all([
-      getFormatData(filePath),
-      getAllStreams(filePath),
+    const { stdout } = await runFfprobe([
+      '-of', 'json', '-show_chapters', '-show_format', '-show_entries', 'stream', '-i', filePath, '-hide_banner',
     ]);
-    const fileFormat = await getDefaultOutFormat(filePath, formatData);
-    const { streams } = allStreamsResponse;
-    // console.log(streams, formatData, fileFormat);
-    return { formatData, fileFormat, streams };
+
+    let parsedJson;
+    try {
+      // https://github.com/mifi/lossless-cut/issues/1342
+      parsedJson = JSON.parse(stdout);
+    } catch (err) {
+      console.log('ffprobe stdout', stdout);
+      throw new Error('ffprobe returned malformed data');
+    }
+    const { streams = [], format = {}, chapters = [] } = parsedJson;
+    return { format, streams, chapters };
   } catch (err) {
     // Windows will throw error with code ENOENT if format detection fails.
-    if (err.exitCode === 1 || (isWindows && err.code === 'ENOENT')) {
+    if (isExecaFailure(err)) {
       const err2 = new Error(`Unsupported file: ${err.message}`);
       err2.code = 'LLC_FFPROBE_UNSUPPORTED_FILE';
       throw err2;
@@ -285,44 +419,64 @@ export async function readFileMeta(filePath) {
   }
 }
 
-
-function getPreferredCodecFormat({ codec_name: codec, codec_type: type }) {
+function getPreferredCodecFormat(stream) {
   const map = {
-    mp3: 'mp3',
-    opus: 'opus',
-    vorbis: 'ogg',
-    h264: 'mp4',
-    hevc: 'mp4',
-    eac3: 'eac3',
+    mp3: { format: 'mp3', ext: 'mp3' },
+    opus: { format: 'opus', ext: 'opus' },
+    vorbis: { format: 'ogg', ext: 'ogg' },
+    h264: { format: 'mp4', ext: 'mp4' },
+    hevc: { format: 'mp4', ext: 'mp4' },
+    eac3: { format: 'eac3', ext: 'eac3' },
 
-    subrip: 'srt',
+    subrip: { format: 'srt', ext: 'srt' },
+    mov_text: { format: 'mp4', ext: 'mp4' },
 
-    // See mapFormat
-    m4a: 'ipod',
-    aac: 'ipod',
+    m4a: { format: 'ipod', ext: 'm4a' },
+    aac: { format: 'adts', ext: 'aac' },
+    jpeg: { format: 'image2', ext: 'jpeg' },
+    png: { format: 'image2', ext: 'png' },
 
     // TODO add more
     // TODO allow user to change?
   };
 
-  const format = map[codec];
-  if (format) return { ext: getExtensionForFormat(format), format };
-  if (type === 'video') return { ext: 'mkv', format: 'matroska' };
-  if (type === 'audio') return { ext: 'mka', format: 'matroska' };
-  if (type === 'subtitle') return { ext: 'mks', format: 'matroska' };
-  if (type === 'data') return { ext: 'bin', format: 'data' }; // https://superuser.com/questions/1243257/save-data-stream
+  const match = map[stream.codec_name];
+  if (match) return match;
+
+  // default fallbacks:
+  if (stream.codec_type === 'video') return { ext: 'mkv', format: 'matroska' };
+  if (stream.codec_type === 'audio') return { ext: 'mka', format: 'matroska' };
+  if (stream.codec_type === 'subtitle') return { ext: 'mks', format: 'matroska' };
+  if (stream.codec_type === 'data') return { ext: 'bin', format: 'data' }; // https://superuser.com/questions/1243257/save-data-stream
 
   return undefined;
 }
 
-async function extractNonAttachmentStreams({ customOutDir, filePath, streams }) {
-  if (streams.length === 0) return;
+async function extractNonAttachmentStreams({ customOutDir, filePath, streams, enableOverwriteOutput }) {
+  if (streams.length === 0) return [];
 
-  console.log('Extracting', streams.length, 'normal streams');
+  const outStreams = streams.map((s) => ({
+    index: s.index,
+    codec: s.codec_name || s.codec_tag_string || s.codec_type,
+    type: s.codec_type,
+    format: getPreferredCodecFormat(s),
+  }))
+    .filter(({ format, index }) => format != null && index != null);
 
-  const streamArgs = flatMap(streams, ({ index, codec, type, format: { format, ext } }) => [
-    '-map', `0:${index}`, '-c', 'copy', '-f', format, '-y', getOutPath(customOutDir, filePath, `stream-${index}-${type}-${codec}.${ext}`),
-  ]);
+  // console.log(outStreams);
+
+
+  let streamArgs = [];
+  const outPaths = await pMap(outStreams, async ({ index, codec, type, format: { format, ext } }) => {
+    const outPath = getSuffixedOutPath({ customOutDir, filePath, nameSuffix: `stream-${index}-${type}-${codec}.${ext}` });
+    if (!enableOverwriteOutput && await pathExists(outPath)) throw new RefuseOverwriteError();
+
+    streamArgs = [
+      ...streamArgs,
+      '-map', `0:${index}`, '-c', 'copy', '-f', format, '-y', outPath,
+    ];
+    return outPath;
+  }, { concurrency: 1 });
 
   const ffmpegArgs = [
     '-hide_banner',
@@ -333,19 +487,27 @@ async function extractNonAttachmentStreams({ customOutDir, filePath, streams }) 
 
   const { stdout } = await runFfmpeg(ffmpegArgs);
   console.log(stdout);
+
+  return outPaths;
 }
 
-async function extractAttachmentStreams({ customOutDir, filePath, streams }) {
-  if (streams.length === 0) return;
+async function extractAttachmentStreams({ customOutDir, filePath, streams, enableOverwriteOutput }) {
+  if (streams.length === 0) return [];
 
   console.log('Extracting', streams.length, 'attachment streams');
 
-  const streamArgs = flatMap(streams, ({ index, codec_name: codec, codec_type: type }) => {
+  let streamArgs = [];
+  const outPaths = await pMap(streams, async ({ index, codec_name: codec, codec_type: type }) => {
     const ext = codec || 'bin';
-    return [
-      `-dump_attachment:${index}`, getOutPath(customOutDir, filePath, `stream-${index}-${type}-${codec}.${ext}`),
+    const outPath = getSuffixedOutPath({ customOutDir, filePath, nameSuffix: `stream-${index}-${type}-${codec}.${ext}` });
+    if (!enableOverwriteOutput && await pathExists(outPath)) throw new RefuseOverwriteError();
+
+    streamArgs = [
+      ...streamArgs,
+      `-dump_attachment:${index}`, outPath,
     ];
-  });
+    return outPath;
+  }, { concurrency: 1 });
 
   const ffmpegArgs = [
     '-y',
@@ -361,31 +523,24 @@ async function extractAttachmentStreams({ customOutDir, filePath, streams }) {
   } catch (err) {
     // Unfortunately ffmpeg will exit with code 1 even though it's a success
     // Note: This is kind of hacky:
-    if (err.exitCode === 1 && typeof err.stderr === 'string' && err.stderr.includes('At least one output file must be specified')) return;
+    if (err.exitCode === 1 && typeof err.stderr === 'string' && err.stderr.includes('At least one output file must be specified')) return outPaths;
     throw err;
   }
+  return outPaths;
 }
 
 // https://stackoverflow.com/questions/32922226/extract-every-audio-and-subtitles-from-a-video-with-ffmpeg
-export async function extractStreams({ filePath, customOutDir, streams }) {
+export async function extractStreams({ filePath, customOutDir, streams, enableOverwriteOutput }) {
   const attachmentStreams = streams.filter((s) => s.codec_type === 'attachment');
   const nonAttachmentStreams = streams.filter((s) => s.codec_type !== 'attachment');
-
-  const outStreams = nonAttachmentStreams.map((s) => ({
-    index: s.index,
-    codec: s.codec_name || s.codec_tag_string || s.codec_type,
-    type: s.codec_type,
-    format: getPreferredCodecFormat(s),
-  }))
-    .filter(it => it && it.format && it.index != null);
-
-  // console.log(outStreams);
 
   // TODO progress
 
   // Attachment streams are handled differently from normal streams
-  await extractNonAttachmentStreams({ customOutDir, filePath, streams: outStreams });
-  await extractAttachmentStreams({ customOutDir, filePath, streams: attachmentStreams });
+  return [
+    ...(await extractNonAttachmentStreams({ customOutDir, filePath, streams: nonAttachmentStreams, enableOverwriteOutput })),
+    ...(await extractAttachmentStreams({ customOutDir, filePath, streams: attachmentStreams, enableOverwriteOutput })),
+  ];
 }
 
 async function renderThumbnail(filePath, timestamp) {
@@ -399,8 +554,7 @@ async function renderThumbnail(filePath, timestamp) {
     '-',
   ];
 
-  const ffmpegPath = await getFfmpegPath();
-  const { stdout } = await execa(ffmpegPath, args, { encoding: null });
+  const { stdout } = await runFfmpeg(args, { encoding: null });
 
   const blob = new Blob([stdout], { type: 'image/jpeg' });
   return URL.createObjectURL(blob);
@@ -415,8 +569,7 @@ export async function extractSubtitleTrack(filePath, streamId) {
     '-',
   ];
 
-  const ffmpegPath = await getFfmpegPath();
-  const { stdout } = await execa(ffmpegPath, args, { encoding: null });
+  const { stdout } = await runFfmpeg(args, { encoding: null });
 
   const blob = new Blob([stdout], { type: 'text/vtt' });
   return URL.createObjectURL(blob);
@@ -443,14 +596,12 @@ export async function renderThumbnails({ filePath, from, duration, onThumbnail }
 }
 
 
-export async function renderWaveformPng({ filePath, aroundTime, window, color }) {
-  const { from, to } = getIntervalAroundTime(aroundTime, window);
-
+export async function renderWaveformPng({ filePath, start, duration, color }) {
   const args1 = [
     '-hide_banner',
     '-i', filePath,
-    '-ss', from,
-    '-t', to - from,
+    '-ss', start,
+    '-t', duration,
     '-c', 'copy',
     '-vn',
     '-map', 'a:0',
@@ -461,7 +612,7 @@ export async function renderWaveformPng({ filePath, aroundTime, window, color })
   const args2 = [
     '-hide_banner',
     '-i', '-',
-    '-filter_complex', `aformat=channel_layouts=mono,showwavespic=s=640x120:scale=sqrt:colors=${color}`,
+    '-filter_complex', `showwavespic=s=2000x300:scale=lin:filter=peak:split_channels=1:colors=${color}`,
     '-frames:v', '1',
     '-vcodec', 'png',
     '-f', 'image2',
@@ -469,31 +620,152 @@ export async function renderWaveformPng({ filePath, aroundTime, window, color })
   ];
 
   console.log(getFfCommandLine('ffmpeg1', args1));
-  console.log(getFfCommandLine('ffmpeg2', args2));
+  console.log('|', getFfCommandLine('ffmpeg2', args2));
 
   let ps1;
   let ps2;
   try {
-    const ffmpegPath = getFfmpegPath();
-    ps1 = execa(ffmpegPath, args1, { encoding: null, buffer: false });
-    ps2 = execa(ffmpegPath, args2, { encoding: null });
+    ps1 = runFfmpeg(args1, { encoding: null, buffer: false }, { logCli: false });
+    ps2 = runFfmpeg(args2, { encoding: null }, { logCli: false });
     ps1.stdout.pipe(ps2.stdin);
 
-    const { stdout } = await ps2;
+    const timer = setTimeout(() => {
+      ps1.kill();
+      ps2.kill();
+      console.warn('ffmpeg timed out');
+    }, 10000);
+
+    let stdout;
+    try {
+      ({ stdout } = await ps2);
+    } finally {
+      clearTimeout(timer);
+    }
 
     const blob = new Blob([stdout], { type: 'image/png' });
 
     return {
       url: URL.createObjectURL(blob),
-      from,
-      aroundTime,
-      to,
+      from: start,
+      to: start + duration,
+      duration,
+      createdAt: new Date(),
     };
   } catch (err) {
     if (ps1) ps1.kill();
     if (ps2) ps2.kill();
     throw err;
   }
+}
+
+const getInputSeekArgs = ({ filePath, from, to }) => [
+  ...(from != null ? ['-ss', from.toFixed(5)] : []),
+  '-i', filePath,
+  ...(to != null ? ['-t', (to - from).toFixed(5)] : []),
+];
+
+const getSegmentOffset = (from) => (from != null ? from : 0);
+
+function adjustSegmentsWithOffset({ segments, from }) {
+  const offset = getSegmentOffset(from);
+  return segments.map(({ start, end }) => ({ start: start + offset, end: end != null ? end + offset : end }));
+}
+
+export function mapTimesToSegments(times) {
+  const segments = [];
+  for (let i = 0; i < times.length; i += 1) {
+    const start = times[i];
+    const end = times[i + 1];
+    if (start != null) segments.push({ start, end }); // end undefined is allowed (means until end of video)
+  }
+  return segments;
+}
+
+// https://stackoverflow.com/questions/35675529/using-ffmpeg-how-to-do-a-scene-change-detection-with-timecode
+export async function detectSceneChanges({ filePath, minChange, onProgress, from, to }) {
+  const args = [
+    '-hide_banner',
+    ...getInputSeekArgs({ filePath, from, to }),
+    '-filter_complex', `select='gt(scene,${minChange})',metadata=print:file=-`,
+    '-f', 'null', '-',
+  ];
+  const process = runFfmpeg(args, { encoding: null, buffer: false });
+
+  const times = [0];
+
+  handleProgress(process, to - from, onProgress);
+  const rl = readline.createInterface({ input: process.stdout });
+  rl.on('line', (line) => {
+    const match = line.match(/^frame:\d+\s+pts:\d+\s+pts_time:([\d.]+)/);
+    if (!match) return;
+    const time = parseFloat(match[1]);
+    if (Number.isNaN(time) || time <= times[times.length - 1]) return;
+    times.push(time);
+  });
+
+  await process;
+
+  const segments = mapTimesToSegments(times);
+
+  return adjustSegmentsWithOffset({ segments, from });
+}
+
+
+export async function detectIntervals({ filePath, customArgs, onProgress, from, to, matchLineTokens }) {
+  const args = [
+    '-hide_banner',
+    ...getInputSeekArgs({ filePath, from, to }),
+    ...customArgs,
+    '-f', 'null', '-',
+  ];
+  const process = runFfmpeg(args, { encoding: null, buffer: false });
+
+  const segments = [];
+
+  function customMatcher(line) {
+    const { start: startStr, end: endStr } = matchLineTokens(line);
+    const start = parseFloat(startStr);
+    const end = parseFloat(endStr);
+    if (start == null || end == null || Number.isNaN(start) || Number.isNaN(end)) return;
+    segments.push({ start, end });
+  }
+  handleProgress(process, to - from, onProgress, customMatcher);
+
+  await process;
+  return adjustSegmentsWithOffset({ segments, from });
+}
+
+const mapFilterOptions = (options) => Object.entries(options).map(([key, value]) => `${key}=${value}`).join(':');
+
+export async function blackDetect({ filePath, filterOptions, onProgress, from, to }) {
+  function matchLineTokens(line) {
+    const match = line.match(/^[blackdetect\s*@\s*0x[0-9a-f]+] black_start:([\d\\.]+) black_end:([\d\\.]+) black_duration:[\d\\.]+/);
+    if (!match) return {};
+    return {
+      start: parseFloat(match[1]),
+      end: parseFloat(match[2]),
+    };
+  }
+  const customArgs = ['-vf', `blackdetect=${mapFilterOptions(filterOptions)}`, '-an'];
+  return detectIntervals({ filePath, onProgress, from, to, matchLineTokens, customArgs });
+}
+
+export async function silenceDetect({ filePath, filterOptions, onProgress, from, to }) {
+  function matchLineTokens(line) {
+    const match = line.match(/^[silencedetect\s*@\s*0x[0-9a-f]+] silence_end: ([\d\\.]+)[|\s]+silence_duration: ([\d\\.]+)/);
+    if (!match) return {};
+    const end = parseFloat(match[1]);
+    const silenceDuration = parseFloat(match[2]);
+    if (Number.isNaN(end) || Number.isNaN(silenceDuration)) return {};
+    const start = end - silenceDuration;
+    if (start < 0 || end <= 0 || start >= end) return {};
+    return {
+      start,
+      end,
+    };
+  }
+  const customArgs = ['-af', `silencedetect=${mapFilterOptions(filterOptions)}`, '-vn'];
+  return detectIntervals({ filePath, onProgress, from, to, matchLineTokens, customArgs });
 }
 
 export async function extractWaveform({ filePath, outPath }) {
@@ -523,41 +795,44 @@ export async function extractWaveform({ filePath, outPath }) {
   console.timeEnd('ffmpeg');
 }
 
-// See also capture-frame.js
-export async function captureFrame({ timestamp, videoPath, outPath }) {
-  const args = [
+function getFffmpegJpegQuality(quality) {
+  // Normal range for JPEG is 2-31 with 31 being the worst quality.
+  const qMin = 2;
+  const qMax = 31;
+  return Math.min(Math.max(qMin, quality, Math.round((1 - quality) * (qMax - qMin) + qMin)), qMax);
+}
+
+export async function captureFrame({ timestamp, videoPath, outPath, quality }) {
+  const ffmpegQuality = getFffmpegJpegQuality(quality);
+  await runFfmpeg([
     '-ss', timestamp,
     '-i', videoPath,
     '-vframes', '1',
-    '-q:v', '3',
+    '-q:v', ffmpegQuality,
     '-y', outPath,
+  ]);
+}
+
+export async function captureFrames({ from, to, videoPath, outPathTemplate, quality, filter, framePts, onProgress }) {
+  const ffmpegQuality = getFffmpegJpegQuality(quality);
+
+  const args = [
+    '-ss', from,
+    '-i', videoPath,
+    '-t', Math.max(0, to - from),
+    '-q:v', ffmpegQuality,
+    ...(filter != null ? ['-vf', filter] : []),
+    // https://superuser.com/questions/1336285/use-ffmpeg-for-thumbnail-selections
+    ...(framePts ? ['-frame_pts', '1'] : []),
+    '-vsync', '0', // else we get a ton of duplicates (thumbnail filter)
+    '-y', outPathTemplate,
   ];
 
-  const ffmpegPath = getFfmpegPath();
-  await execa(ffmpegPath, args, { encoding: null });
-}
+  const process = runFfmpeg(args, { encoding: null, buffer: false });
 
-// https://www.ffmpeg.org/doxygen/3.2/libavutil_2utils_8c_source.html#l00079
-export const defaultProcessedCodecTypes = [
-  'video',
-  'audio',
-  'subtitle',
-  'attachment',
-];
+  handleProgress(process, to - from, onProgress);
 
-export const isMov = (format) => ['ismv', 'ipod', 'mp4', 'mov'].includes(format);
-
-export function isStreamThumbnail(stream) {
-  return stream && stream.disposition && stream.disposition.attached_pic === 1;
-}
-
-const getAudioStreams = (streams) => streams.filter(stream => stream.codec_type === 'audio');
-
-export function isAudioDefinitelyNotSupported(streams) {
-  const audioStreams = getAudioStreams(streams);
-  if (audioStreams.length === 0) return false;
-  // TODO this could be improved
-  return audioStreams.every(stream => ['ac3'].includes(stream.codec_name));
+  await process;
 }
 
 export function isIphoneHevc(format, streams) {
@@ -565,6 +840,12 @@ export function isIphoneHevc(format, streams) {
   const makeTag = format.tags && format.tags['com.apple.quicktime.make'];
   const modelTag = format.tags && format.tags['com.apple.quicktime.model'];
   return (makeTag === 'Apple' && modelTag.startsWith('iPhone'));
+}
+
+export function isProblematicAvc1(outFormat, streams) {
+  // it seems like this only happens for files that are also 4.2.2 10bit (yuv422p10le)
+  // https://trac.ffmpeg.org/wiki/Chroma%20Subsampling
+  return isMov(outFormat) && streams.some((s) => s.codec_name === 'h264' && s.codec_tag === '0x31637661' && s.codec_tag_string === 'avc1' && s.pix_fmt === 'yuv422p10le');
 }
 
 export function getStreamFps(stream) {
@@ -617,7 +898,7 @@ function createRawFfmpeg({ fps = 25, path, inWidth, inHeight, seekTo, oneFrameOn
   // console.log(args);
 
   return {
-    process: execa(getFfmpegPath(), args, execaOpts),
+    process: runFfmpeg(args, execaOpts, { logCli: false }),
     width: newWidth,
     height: newHeight,
     channels: 4,
@@ -666,4 +947,166 @@ export function getTimecodeFromStreams(streams) {
     }
   });
   return foundTimecode;
+}
+
+export async function runFfmpegStartupCheck() {
+  // will throw if exit code != 0
+  await runFfmpeg(['-hide_banner', '-f', 'lavfi', '-i', 'nullsrc=s=256x256:d=1', '-f', 'null', '-']);
+}
+
+export async function html5ify({ outPath, filePath: filePathArg, speed, hasAudio, hasVideo, onProgress }) {
+  let audio;
+  if (hasAudio) {
+    if (speed === 'slowest') audio = 'hq';
+    else if (['slow-audio', 'fast-audio', 'fastest-audio'].includes(speed)) audio = 'lq';
+    else if (['fast-audio-remux', 'fastest-audio-remux'].includes(speed)) audio = 'copy';
+  }
+
+  let video;
+  if (hasVideo) {
+    if (speed === 'slowest') video = 'hq';
+    else if (['slow-audio', 'slow'].includes(speed)) video = 'lq';
+    else video = 'copy';
+  }
+
+  console.log('Making HTML5 friendly version', { filePathArg, outPath, video, audio });
+
+  let videoArgs;
+  let audioArgs;
+
+  // h264/aac_at: No licensing when using HW encoder (Video/Audio Toolbox on Mac)
+  // https://github.com/mifi/lossless-cut/issues/372#issuecomment-810766512
+
+  const targetHeight = 400;
+
+  switch (video) {
+    case 'hq': {
+      if (isMac) {
+        videoArgs = ['-vf', 'format=yuv420p', '-allow_sw', '1', '-vcodec', 'h264', '-b:v', '15M'];
+      } else {
+        // AV1 is very slow
+        // videoArgs = ['-vf', 'format=yuv420p', '-sws_flags', 'neighbor', '-vcodec', 'libaom-av1', '-crf', '30', '-cpu-used', '8'];
+        // Theora is a bit faster but not that much
+        // videoArgs = ['-vf', '-c:v', 'libtheora', '-qscale:v', '1'];
+        // videoArgs = ['-vf', 'format=yuv420p', '-c:v', 'libvpx-vp9', '-crf', '30', '-b:v', '0', '-row-mt', '1'];
+        // x264 can only be used in GPL projects
+        videoArgs = ['-vf', 'format=yuv420p', '-c:v', 'libx264', '-profile:v', 'high', '-preset:v', 'slow', '-crf', '17'];
+      }
+      break;
+    }
+    case 'lq': {
+      if (isMac) {
+        videoArgs = ['-vf', `scale=-2:${targetHeight},format=yuv420p`, '-allow_sw', '1', '-sws_flags', 'lanczos', '-vcodec', 'h264', '-b:v', '1500k'];
+      } else {
+        // videoArgs = ['-vf', `scale=-2:${targetHeight},format=yuv420p`, '-sws_flags', 'neighbor', '-c:v', 'libtheora', '-qscale:v', '1'];
+        // x264 can only be used in GPL projects
+        videoArgs = ['-vf', `scale=-2:${targetHeight},format=yuv420p`, '-sws_flags', 'neighbor', '-c:v', 'libx264', '-profile:v', 'baseline', '-x264opts', 'level=3.0', '-preset:v', 'ultrafast', '-crf', '28'];
+      }
+      break;
+    }
+    case 'copy': {
+      videoArgs = ['-vcodec', 'copy'];
+      break;
+    }
+    default: {
+      videoArgs = ['-vn'];
+    }
+  }
+
+  switch (audio) {
+    case 'hq': {
+      if (isMac) {
+        audioArgs = ['-acodec', 'aac_at', '-b:a', '192k'];
+      } else {
+        audioArgs = ['-acodec', 'flac'];
+      }
+      break;
+    }
+    case 'lq': {
+      if (isMac) {
+        audioArgs = ['-acodec', 'aac_at', '-ar', '44100', '-ac', '2', '-b:a', '96k'];
+      } else {
+        audioArgs = ['-acodec', 'flac', '-ar', '11025', '-ac', '2'];
+      }
+      break;
+    }
+    case 'copy': {
+      audioArgs = ['-acodec', 'copy'];
+      break;
+    }
+    default: {
+      audioArgs = ['-an'];
+    }
+  }
+
+  const ffmpegArgs = [
+    '-hide_banner',
+
+    '-i', filePathArg,
+    ...videoArgs,
+    ...audioArgs,
+    '-sn',
+    '-y', outPath,
+  ];
+
+  const duration = await getDuration(filePathArg);
+  const process = runFfmpeg(ffmpegArgs);
+  if (duration) handleProgress(process, duration, onProgress);
+
+  const { stdout } = await process;
+  console.log(stdout);
+}
+
+// https://superuser.com/questions/543589/information-about-ffmpeg-command-line-options
+export const getExperimentalArgs = (ffmpegExperimental) => (ffmpegExperimental ? ['-strict', 'experimental'] : []);
+
+export const getVideoTimescaleArgs = (videoTimebase) => (videoTimebase != null ? ['-video_track_timescale', videoTimebase] : []);
+
+// inspired by https://gist.github.com/fernandoherreradelasheras/5eca67f4200f1a7cc8281747da08496e
+export async function cutEncodeSmartPart({ filePath, cutFrom, cutTo, outPath, outFormat, videoCodec, videoBitrate, videoTimebase, allFilesMeta, copyFileStreams, videoStreamIndex, ffmpegExperimental }) {
+  function getVideoArgs({ streamIndex, outputIndex }) {
+    if (streamIndex !== videoStreamIndex) return undefined;
+
+    const args = [
+      `-c:${outputIndex}`, videoCodec,
+      `-b:${outputIndex}`, videoBitrate,
+    ];
+
+    // seems like ffmpeg handles this itself well when encoding same source file
+    // if (videoLevel != null) args.push(`-level:${outputIndex}`, videoLevel);
+    // if (videoProfile != null) args.push(`-profile:${outputIndex}`, videoProfile);
+
+    return args;
+  }
+
+  const mapStreamsArgs = getMapStreamsArgs({
+    allFilesMeta,
+    copyFileStreams,
+    outFormat,
+    getVideoArgs,
+  });
+
+  const ffmpegArgs = [
+    '-hide_banner',
+    // No progress if we set loglevel warning :(
+    // '-loglevel', 'warning',
+
+    '-ss', cutFrom.toFixed(5), // if we don't -ss before -i, seeking will be slow for long files, see https://github.com/mifi/lossless-cut/issues/126#issuecomment-1135451043
+    '-i', filePath,
+    '-ss', '0', // If we don't do this, the output seems to start with an empty black after merging with the encoded part
+    '-t', (cutTo - cutFrom).toFixed(5),
+
+    ...mapStreamsArgs,
+
+    // See https://github.com/mifi/lossless-cut/issues/170
+    '-ignore_unknown',
+
+    ...getVideoTimescaleArgs(videoTimebase),
+
+    ...getExperimentalArgs(ffmpegExperimental),
+
+    '-f', outFormat, '-y', outPath,
+  ];
+
+  await runFfmpeg(ffmpegArgs);
 }
