@@ -1,36 +1,36 @@
 import { useCallback } from 'react';
 import flatMap from 'lodash/flatMap';
-import flatMapDeep from 'lodash/flatMapDeep';
 import sum from 'lodash/sum';
 import pMap from 'p-map';
 
-import { getOutPath, transferTimestamps, getOutFileExtension, getOutDir, isMac } from '../util';
-import { isCuttingStart, isCuttingEnd, handleProgress, getFfCommandLine, getFfmpegPath, getDuration, runFfmpeg, createChaptersFromSegments } from '../ffmpeg';
+import { getSuffixedOutPath, transferTimestamps, getOutFileExtension, getOutDir, deleteDispositionValue, getHtml5ifiedPath } from '../util';
+import { isCuttingStart, isCuttingEnd, handleProgress, getFfCommandLine, getDuration, runFfmpeg, createChaptersFromSegments, readFileMeta, cutEncodeSmartPart, getExperimentalArgs, html5ify as ffmpegHtml5ify, getVideoTimescaleArgs, RefuseOverwriteError, logStdoutStderr } from '../ffmpeg';
+import { getMapStreamsArgs, getStreamIdsToCopy } from '../util/streams';
+import { getSmartCutParams } from '../smartcut';
 
-const execa = window.require('execa');
-const { join } = window.require('path');
-const fs = window.require('fs-extra');
+const { join, resolve, dirname } = window.require('path');
+const { pathExists } = window.require('fs-extra');
+const { writeFile, unlink, mkdir } = window.require('fs/promises');
 const stringToStream = window.require('string-to-stream');
 
 async function writeChaptersFfmetadata(outDir, chapters) {
-  if (!chapters) return undefined;
+  if (!chapters || chapters.length === 0) return undefined;
 
   const path = join(outDir, `ffmetadata-${new Date().getTime()}.txt`);
 
-  const ffmetadata = chapters.map(({ start, end, name }, i) => {
-    const nameOut = name || `Chapter ${i + 1}`;
-    return `[CHAPTER]\nTIMEBASE=1/1000\nSTART=${Math.floor(start * 1000)}\nEND=${Math.floor(end * 1000)}\ntitle=${nameOut}`;
-  }).join('\n\n');
-  // console.log(ffmetadata);
-  await fs.writeFile(path, ffmetadata);
+  const ffmetadata = chapters.map(({ start, end, name }) => (
+    `[CHAPTER]\nTIMEBASE=1/1000\nSTART=${Math.floor(start * 1000)}\nEND=${Math.floor(end * 1000)}\ntitle=${name || ''}`
+  )).join('\n\n');
+  console.log('Writing chapters', ffmetadata);
+  await writeFile(path, ffmetadata);
   return path;
 }
 
 function getMovFlags({ preserveMovData, movFastStart }) {
   const flags = [];
 
-  // https://video.stackexchange.com/questions/23741/how-to-prevent-ffmpeg-from-dropping-metadata
   // https://video.stackexchange.com/a/26084/29486
+  // https://github.com/mifi/lossless-cut/issues/331#issuecomment-623401794
   if (preserveMovData) flags.push('use_metadata_tags');
 
   // https://github.com/mifi/lossless-cut/issues/347
@@ -40,142 +40,294 @@ function getMovFlags({ preserveMovData, movFastStart }) {
   return flatMap(flags, flag => ['-movflags', flag]);
 }
 
-function useFfmpegOperations({ filePath, enableTransferTimestamps }) {
+function getMatroskaFlags() {
+  return [
+    '-default_mode', 'infer_no_subs',
+    // because it makes sense to not force subtitles disposition to "default" if they were not default in the input file
+    // after some testing, it seems that default is actually "infer", contrary to what is documented (ffmpeg doc says "passthrough" is default)
+    // https://ffmpeg.org/ffmpeg-formats.html#Options-8
+    // https://github.com/mifi/lossless-cut/issues/972#issuecomment-1015176316
+  ];
+}
+
+const getChaptersInputArgs = (ffmetadataPath) => (ffmetadataPath ? ['-f', 'ffmetadata', '-i', ffmetadataPath] : []);
+
+const tryDeleteFiles = async (paths) => pMap(paths, (path) => {
+  unlink(path).catch((err) => console.error('Failed to delete', path, err));
+}, { concurrency: 5 });
+
+function useFfmpegOperations({ filePath, enableTransferTimestamps, needSmartCut }) {
   const optionalTransferTimestamps = useCallback(async (...args) => {
     if (enableTransferTimestamps) await transferTimestamps(...args);
   }, [enableTransferTimestamps]);
 
-  // const cut = useCallback(, [filePath, optionalTransferTimestamps]);
+  const concatFiles = useCallback(async ({ paths, outDir, outPath, metadataFromPath, includeAllStreams, streams, outFormat, ffmpegExperimental, onProgress = () => {}, preserveMovData, movFastStart, chapters, preserveMetadataOnMerge, videoTimebase, appendFfmpegCommandLog }) => {
+    console.log('Merging files', { paths }, 'to', outPath);
 
-  const cutMultiple = useCallback(async ({
-    outputDir, segments, segmentsFileNames, videoDuration, rotation,
-    onProgress: onTotalProgress, keyframeCut, copyFileStreams, outFormat,
-    appendFfmpegCommandLog, shortestFlag, ffmpegExperimental, preserveMovData, movFastStart, avoidNegativeTs,
-    customTagsByFile, customTagsByStreamId, dispositionByStreamId,
-  }) => {
-    async function cutSingle({ cutFrom, cutTo, onProgress, outPath }) {
-      const cuttingStart = isCuttingStart(cutFrom);
-      const cuttingEnd = isCuttingEnd(cutTo, videoDuration);
-      console.log('Exporting from', cuttingStart ? cutFrom : 'start', 'to', cuttingEnd ? cutTo : 'end');
+    const durations = await pMap(paths, getDuration, { concurrency: 1 });
+    const totalDuration = sum(durations);
 
-      const ssBeforeInput = keyframeCut;
+    let chaptersPath;
+    if (chapters) {
+      const chaptersWithNames = chapters.map((chapter, i) => ({ ...chapter, name: chapter.name || `Chapter ${i + 1}` }));
+      chaptersPath = await writeChaptersFfmetadata(outDir, chaptersWithNames);
+    }
 
-      const cutDuration = cutTo - cutFrom;
+    try {
+      let inputArgs = [];
+      let inputIndex = 0;
 
-      // Don't cut if no need: https://github.com/mifi/lossless-cut/issues/50
-      const cutFromArgs = cuttingStart ? ['-ss', cutFrom.toFixed(5)] : [];
-      const cutToArgs = cuttingEnd ? ['-t', cutDuration.toFixed(5)] : [];
-
-      const copyFileStreamsFiltered = copyFileStreams.filter(({ streamIds }) => streamIds.length > 0);
-
-      // remove -avoid_negative_ts make_zero when not cutting start (no -ss), or else some videos get blank first frame in QuickLook
-      const avoidNegativeTsArgs = cuttingStart && avoidNegativeTs ? ['-avoid_negative_ts', avoidNegativeTs] : [];
-
-      const inputArgs = flatMap(copyFileStreamsFiltered, ({ path }) => ['-i', path]);
-      const inputCutArgs = ssBeforeInput ? [
-        ...cutFromArgs,
-        ...inputArgs,
-        ...cutToArgs,
-        ...avoidNegativeTsArgs,
-      ] : [
-        ...inputArgs,
-        ...cutFromArgs,
-        ...cutToArgs,
-      ];
-
-      const rotationArgs = rotation !== undefined ? ['-metadata:s:v:0', `rotate=${360 - rotation}`] : [];
-
-      // This function tries to calculate the output stream index needed for -metadata:s:x and -disposition:x arguments
-      // It is based on the assumption that copyFileStreamsFiltered contains the order of the input files (and their respective streams orders) sent to ffmpeg, to hopefully calculate the same output stream index values that ffmpeg does internally.
-      // It also takes into account previously added files that have been removed and disabled streams.
-      function mapInputStreamIndexToOutputIndex(inputFilePath, inputFileStreamIndex) {
-        let streamCount = 0;
-        // Count copied streams of all files until this input file
-        const foundFile = copyFileStreamsFiltered.find(({ path: path2, streamIds }) => {
-          if (path2 === inputFilePath) return true;
-          streamCount += streamIds.length;
-          return false;
-        });
-        if (!foundFile) return undefined; // Could happen if a tag has been edited on an external file, then the file was removed
-
-        // Then add the index of the current stream index to the count
-        const copiedStreamIndex = foundFile.streamIds.indexOf(String(inputFileStreamIndex));
-        if (copiedStreamIndex === -1) return undefined; // Could happen if a tag has been edited on a stream, but the stream is disabled
-        return streamCount + copiedStreamIndex;
+      // Keep track of input index to be used later
+      // eslint-disable-next-line no-inner-declarations
+      function addInput(args) {
+        inputArgs = [...inputArgs, ...args];
+        const retIndex = inputIndex;
+        inputIndex += 1;
+        return retIndex;
       }
 
-      // The structure is deep! file -> stream -> key -> value Example: { 'file.mp4': { 0: { key: 'value' } } }
-      const deepMap = (root, fn) => flatMapDeep(
-        Object.entries(root), ([path, streamsMap]) => (
-          Object.entries(streamsMap || {}).map(([streamId, tagsMap]) => (
-            Object.entries(tagsMap || {}).map(([key, value]) => fn(path, streamId, key, value))))),
-      );
+      // concat list - always first
+      addInput([
+        // https://blog.yo1.dog/fix-for-ffmpeg-protocol-not-on-whitelist-error-for-urls/
+        '-f', 'concat', '-safe', '0', '-protocol_whitelist', 'file,pipe,fd',
+        '-i', '-',
+      ]);
 
-      const customTagsArgs = [
-        // Main file metadata:
-        ...flatMap(Object.entries(customTagsByFile[filePath] || []), ([key, value]) => ['-metadata', `${key}=${value}`]),
+      let metadataSourceIndex;
+      if (preserveMetadataOnMerge) {
+        // If preserve metadata, add the first file (we will get metadata from this input)
+        metadataSourceIndex = addInput(['-i', metadataFromPath]);
+      }
 
-        // Example: { 'file.mp4': { 0: { tag_name: 'Tag Value' } } }
-        ...deepMap(customTagsByStreamId, (path, streamId, tag, value) => {
-          const outputIndex = mapInputStreamIndexToOutputIndex(path, parseInt(streamId, 10));
-          if (outputIndex == null) return [];
-          return [`-metadata:s:${outputIndex}`, `${tag}=${value}`];
-        }),
-      ];
+      let chaptersInputIndex;
+      if (chaptersPath) {
+        // if chapters, add chapters source file
+        chaptersInputIndex = addInput(getChaptersInputArgs(chaptersPath));
+      }
 
-      // Example: { 'file.mp4': { 0: { attached_pic: 1 } } }
-      const customDispositionArgs = deepMap(dispositionByStreamId, (path, streamId, disposition, value) => {
-        if (value !== 1) return [];
-        const outputIndex = mapInputStreamIndexToOutputIndex(path, parseInt(streamId, 10));
-        if (outputIndex == null) return [];
-        return [`-disposition:${outputIndex}`, String(disposition)];
+      const { streamIdsToCopy, excludedStreamIds } = getStreamIdsToCopy({ streams, includeAllStreams });
+      const mapStreamsArgs = getMapStreamsArgs({
+        allFilesMeta: { [metadataFromPath]: { streams } },
+        copyFileStreams: [{ path: metadataFromPath, streamIds: streamIdsToCopy }],
+        outFormat,
+        manuallyCopyDisposition: true,
       });
 
+      // Keep this similar to cutSingle()
       const ffmpegArgs = [
         '-hide_banner',
         // No progress if we set loglevel warning :(
         // '-loglevel', 'warning',
 
-        ...inputCutArgs,
+        ...inputArgs,
 
-        '-c', 'copy',
+        ...mapStreamsArgs,
 
-        ...(shortestFlag ? ['-shortest'] : []),
+        // -map_metadata 0 with concat demuxer doesn't transfer metadata from the concat'ed file input (index 0) when merging.
+        // So we use the first file file (index 1) for metadata
+        // Can only do this if allStreams (-map 0) is set
+        ...(metadataSourceIndex != null ? ['-map_metadata', metadataSourceIndex] : []),
 
-        ...flatMapDeep(copyFileStreamsFiltered, ({ streamIds }, fileIndex) => streamIds.map(streamId => ['-map', `${fileIndex}:${streamId}`])),
-        '-map_metadata', '0',
+        ...(chaptersInputIndex != null ? ['-map_chapters', chaptersInputIndex] : []),
 
         ...getMovFlags({ preserveMovData, movFastStart }),
-
-        ...customTagsArgs,
-
-        ...customDispositionArgs,
+        ...getMatroskaFlags(),
 
         // See https://github.com/mifi/lossless-cut/issues/170
         '-ignore_unknown',
 
-        // https://superuser.com/questions/543589/information-about-ffmpeg-command-line-options
-        ...(ffmpegExperimental ? ['-strict', 'experimental'] : []),
+        ...getExperimentalArgs(ffmpegExperimental),
 
-        ...rotationArgs,
+        ...getVideoTimescaleArgs(videoTimebase),
 
-        '-f', outFormat, '-y', outPath,
+        ...(outFormat ? ['-f', outFormat] : []),
+        '-y', outPath,
       ];
+
+      // https://superuser.com/questions/787064/filename-quoting-in-ffmpeg-concat
+      // Must add "file:" or we get "Impossible to open 'pipe:xyz.mp4'" on newer ffmpeg versions
+      // https://superuser.com/questions/718027/ffmpeg-concat-doesnt-work-with-absolute-path
+      const concatTxt = paths.map(file => `file 'file:${resolve(file).replace(/'/g, "'\\''")}'`).join('\n');
 
       const ffmpegCommandLine = getFfCommandLine('ffmpeg', ffmpegArgs);
 
-      console.log(ffmpegCommandLine);
-      appendFfmpegCommandLog(ffmpegCommandLine);
+      const fullCommandLine = `echo -e "${concatTxt.replace(/\n/, '\\n')}" | ${ffmpegCommandLine}`;
+      console.log(fullCommandLine);
+      appendFfmpegCommandLog(fullCommandLine);
 
-      const ffmpegPath = getFfmpegPath();
-      const process = execa(ffmpegPath, ffmpegArgs);
-      handleProgress(process, cutDuration, onProgress);
+      const process = runFfmpeg(ffmpegArgs);
+
+      handleProgress(process, totalDuration, onProgress);
+
+      stringToStream(concatTxt).pipe(process.stdin);
+
       const result = await process;
-      console.log(result.stdout);
+      logStdoutStderr(result);
 
-      await optionalTransferTimestamps(filePath, outPath, cutFrom);
+      await optionalTransferTimestamps(metadataFromPath, outPath);
+
+      return { haveExcludedStreams: excludedStreamIds.length > 0 };
+    } finally {
+      if (chaptersPath) await tryDeleteFiles([chaptersPath]);
+    }
+  }, [optionalTransferTimestamps]);
+
+  const cutSingle = useCallback(async ({
+    keyframeCut: ssBeforeInput, avoidNegativeTs, copyFileStreams, cutFrom, cutTo, chaptersPath, onProgress, outPath,
+    videoDuration, rotation, allFilesMeta, outFormat, appendFfmpegCommandLog, shortestFlag, ffmpegExperimental, preserveMovData, movFastStart, customTagsByFile, customTagsByStreamId, dispositionByStreamId, videoTimebase,
+  }) => {
+    const cuttingStart = isCuttingStart(cutFrom);
+    const cuttingEnd = isCuttingEnd(cutTo, videoDuration);
+    console.log('Cutting from', cuttingStart ? cutFrom : 'start', 'to', cuttingEnd ? cutTo : 'end');
+
+    const cutDuration = cutTo - cutFrom;
+
+    // Don't cut if no need: https://github.com/mifi/lossless-cut/issues/50
+    const cutFromArgs = cuttingStart ? ['-ss', cutFrom.toFixed(5)] : [];
+    const cutToArgs = cuttingEnd ? ['-t', cutDuration.toFixed(5)] : [];
+
+    const copyFileStreamsFiltered = copyFileStreams.filter(({ streamIds }) => streamIds.length > 0);
+
+    // remove -avoid_negative_ts make_zero when not cutting start (no -ss), or else some videos get blank first frame in QuickLook
+    const avoidNegativeTsArgs = cuttingStart && avoidNegativeTs ? ['-avoid_negative_ts', avoidNegativeTs] : [];
+
+    const inputFilesArgs = flatMap(copyFileStreamsFiltered, ({ path }) => ['-i', path]);
+    const inputFilesArgsWithCuts = ssBeforeInput ? [
+      ...cutFromArgs,
+      ...inputFilesArgs,
+      ...cutToArgs,
+      ...avoidNegativeTsArgs,
+    ] : [
+      ...inputFilesArgs,
+      ...cutFromArgs,
+      ...cutToArgs,
+    ];
+
+    const inputArgs = [
+      ...inputFilesArgsWithCuts,
+      ...getChaptersInputArgs(chaptersPath),
+    ];
+
+    const chaptersInputIndex = copyFileStreamsFiltered.length;
+
+    const rotationArgs = rotation !== undefined ? ['-metadata:s:v:0', `rotate=${360 - rotation}`] : [];
+
+    // This function tries to calculate the output stream index needed for -metadata:s:x and -disposition:x arguments
+    // It is based on the assumption that copyFileStreamsFiltered contains the order of the input files (and their respective streams orders) sent to ffmpeg, to hopefully calculate the same output stream index values that ffmpeg does internally.
+    // It also takes into account previously added files that have been removed and disabled streams.
+    function mapInputStreamIndexToOutputIndex(inputFilePath, inputFileStreamIndex) {
+      let streamCount = 0;
+      // Count copied streams of all files until this input file
+      const foundFile = copyFileStreamsFiltered.find(({ path: path2, streamIds }) => {
+        if (path2 === inputFilePath) return true;
+        streamCount += streamIds.length;
+        return false;
+      });
+      if (!foundFile) return undefined; // Could happen if a tag has been edited on an external file, then the file was removed
+
+      // Then add the index of the current stream index to the count
+      const copiedStreamIndex = foundFile.streamIds.indexOf(inputFileStreamIndex);
+      if (copiedStreamIndex === -1) return undefined; // Could happen if a tag has been edited on a stream, but the stream is disabled
+      return streamCount + copiedStreamIndex;
     }
 
+    function lessDeepMap(root, fn) {
+      let ret = [];
+      Object.entries(root).forEach(([path, streamsMap]) => (
+        Object.entries(streamsMap || {}).forEach(([streamId, value]) => {
+          ret = [...ret, ...fn(path, streamId, value)];
+        })));
+
+      return ret;
+    }
+
+    // The structure is deep! file -> stream -> key -> value Example: { 'file.mp4': { 0: { key: 'value' } } }
+    const deepMap = (root, fn) => lessDeepMap(root, (path, streamId, tagsMap) => {
+      let ret = [];
+      Object.entries(tagsMap || {}).forEach(([key, value]) => {
+        ret = [...ret, ...fn(path, streamId, key, value)];
+      });
+      return ret;
+    });
+
+    const customTagsArgs = [
+      // Main file metadata:
+      ...flatMap(Object.entries(customTagsByFile[filePath] || []), ([key, value]) => ['-metadata', `${key}=${value}`]),
+
+      // Example: { 'file.mp4': { 0: { tag_name: 'Tag Value' } } }
+      ...deepMap(customTagsByStreamId, (path, streamId, tag, value) => {
+        const outputIndex = mapInputStreamIndexToOutputIndex(path, parseInt(streamId, 10));
+        if (outputIndex == null) return [];
+        return [`-metadata:s:${outputIndex}`, `${tag}=${value}`];
+      }),
+    ];
+
+    const mapStreamsArgs = getMapStreamsArgs({ copyFileStreams: copyFileStreamsFiltered, allFilesMeta, outFormat });
+
+    // Example: { 'file.mp4': { 0: { attached_pic: 1 } } }
+    const customDispositionArgs = lessDeepMap(dispositionByStreamId, (path, streamId, disposition) => {
+      if (disposition == null) return [];
+      const outputIndex = mapInputStreamIndexToOutputIndex(path, parseInt(streamId, 10));
+      if (outputIndex == null) return [];
+      // 0 means delete the disposition for this stream
+      const dispositionArg = disposition === deleteDispositionValue ? '0' : disposition;
+      return [`-disposition:${outputIndex}`, String(dispositionArg)];
+    });
+
+    const ffmpegArgs = [
+      '-hide_banner',
+      // No progress if we set loglevel warning :(
+      // '-loglevel', 'warning',
+
+      ...inputArgs,
+
+      ...mapStreamsArgs,
+
+      '-map_metadata', '0',
+
+      ...(chaptersPath ? ['-map_chapters', chaptersInputIndex] : []),
+
+      ...(shortestFlag ? ['-shortest'] : []),
+
+      ...getMovFlags({ preserveMovData, movFastStart }),
+      ...getMatroskaFlags(),
+
+      ...customTagsArgs,
+
+      ...customDispositionArgs,
+
+      // See https://github.com/mifi/lossless-cut/issues/170
+      '-ignore_unknown',
+
+      ...getExperimentalArgs(ffmpegExperimental),
+
+      ...rotationArgs,
+
+      ...getVideoTimescaleArgs(videoTimebase),
+
+      '-f', outFormat, '-y', outPath,
+    ];
+
+    const ffmpegCommandLine = getFfCommandLine('ffmpeg', ffmpegArgs);
+
+    // console.log(ffmpegCommandLine);
+    appendFfmpegCommandLog(ffmpegCommandLine);
+
+    const process = runFfmpeg(ffmpegArgs);
+    handleProgress(process, cutDuration, onProgress);
+    const result = await process;
+    logStdoutStderr(result);
+
+    await optionalTransferTimestamps(filePath, outPath, cutFrom);
+  }, [filePath, optionalTransferTimestamps]);
+
+  const cutMultiple = useCallback(async ({
+    outputDir, customOutDir, segments, outSegFileNames, videoDuration, rotation, detectedFps,
+    onProgress: onTotalProgress, keyframeCut, copyFileStreams, allFilesMeta, outFormat,
+    appendFfmpegCommandLog, shortestFlag, ffmpegExperimental, preserveMovData, movFastStart, avoidNegativeTs,
+    customTagsByFile, customTagsByStreamId, dispositionByStreamId, chapters, preserveMetadataOnMerge,
+    enableOverwriteOutput,
+  }) => {
     console.log('customTagsByFile', customTagsByFile);
     console.log('customTagsByStreamId', customTagsByStreamId);
 
@@ -185,200 +337,145 @@ function useFfmpegOperations({ filePath, enableTransferTimestamps }) {
       return onTotalProgress((sum(Object.values(singleProgresses)) / segments.length));
     }
 
-    const outFiles = [];
+    const chaptersPath = await writeChaptersFfmetadata(outputDir, chapters);
 
-    // eslint-disable-next-line no-restricted-syntax,no-unused-vars
-    for (const [i, { start: cutFrom, end: cutTo }] of segments.entries()) {
-      const fileName = segmentsFileNames[i];
-
-      const outPath = join(outputDir, fileName);
-
-      // eslint-disable-next-line no-await-in-loop
-      await cutSingle({ cutFrom, cutTo, outPath, onProgress: progress => onSingleProgress(i, progress) });
-
-      outFiles.push(outPath);
+    async function checkOverwrite(path) {
+      if (!enableOverwriteOutput && await pathExists(path)) throw new RefuseOverwriteError();
     }
 
-    return outFiles;
-  }, [filePath, optionalTransferTimestamps]);
 
-  const mergeFiles = useCallback(async ({ paths, outDir, outPath, allStreams, outFormat, ffmpegExperimental, onProgress = () => {}, preserveMovData, movFastStart, chapters, preserveMetadataOnMerge }) => {
-    console.log('Merging files', { paths }, 'to', outPath);
+    // This function will either call cutSingle (if no smart cut enabled)
+    // or if enabled, will first cut&encode the part before the next keyframe, trying to match the input file's codec params
+    // then it will cut the part *from* the keyframe to "end", and concat them together and return the concated file
+    // so that for the calling code it looks as if it's just a normal segment
+    async function maybeSmartCutSegment({ start: desiredCutFrom, end: cutTo }, i) {
+      async function makeSegmentOutPath() {
+        const outPath = join(outputDir, outSegFileNames[i]);
+        // because outSegFileNames might contain slashes https://github.com/mifi/lossless-cut/issues/1532
+        const actualOutputDir = dirname(outPath);
+        if (actualOutputDir !== outputDir) await mkdir(actualOutputDir, { recursive: true });
+        return outPath;
+      }
 
-    const durations = await pMap(paths, getDuration, { concurrency: 1 });
-    const totalDuration = sum(durations);
+      if (!needSmartCut) {
+        // old fashioned way
+        const outPath = await makeSegmentOutPath();
+        await checkOverwrite(outPath);
+        await cutSingle({
+          cutFrom: desiredCutFrom, cutTo, chaptersPath, outPath, copyFileStreams, keyframeCut, avoidNegativeTs, videoDuration, rotation, allFilesMeta, outFormat, appendFfmpegCommandLog, shortestFlag, ffmpegExperimental, preserveMovData, movFastStart, customTagsByFile, customTagsByStreamId, dispositionByStreamId, onProgress: (progress) => onSingleProgress(i, progress),
+        });
+        return outPath;
+      }
 
-    const ffmetadataPath = await writeChaptersFfmetadata(outDir, chapters);
+      // smart cut only supports cutting main file (no externally added files)
+      const { streams } = allFilesMeta[filePath];
+      const streamsToCopyFromMainFile = copyFileStreams.find(({ path }) => path === filePath).streamIds
+        .map((streamId) => streams.find((stream) => stream.index === streamId));
+
+      const { cutFrom: encodeCutTo, segmentNeedsSmartCut, videoCodec, videoBitrate, videoStreamIndex, videoTimebase } = await getSmartCutParams({ path: filePath, videoDuration, desiredCutFrom, streams: streamsToCopyFromMainFile });
+
+      if (segmentNeedsSmartCut && !detectedFps) throw new Error('Smart cut is not possible when FPS is unknown');
+
+      console.log('Smart cut on video stream', videoStreamIndex);
+
+      const onCutProgress = (progress) => onSingleProgress(i, progress / 2);
+      const onConcatProgress = (progress) => onSingleProgress(i, (1 + progress) / 2);
+
+      const copyFileStreamsFiltered = [{
+        path: filePath,
+        // with smart cut, we only copy/cut *one* video stream, but *all* other streams (main file only)
+        streamIds: streamsToCopyFromMainFile.filter((stream) => !(stream.codec_type === 'video' && stream.index !== videoStreamIndex)).map((stream) => stream.index),
+      }];
+
+      // eslint-disable-next-line no-shadow
+      const cutEncodeSmartPartWrapper = async ({ cutFrom, cutTo, outPath }) => cutEncodeSmartPart({ filePath, cutFrom, cutTo, outPath, outFormat, videoCodec, videoBitrate, videoStreamIndex, videoTimebase, allFilesMeta, copyFileStreams: copyFileStreamsFiltered, ffmpegExperimental });
+
+      // If we are cutting within two keyframes, just encode the whole part and return that
+      // See https://github.com/mifi/lossless-cut/pull/1267#issuecomment-1236381740
+      if (segmentNeedsSmartCut && encodeCutTo > cutTo) {
+        const outPath = await makeSegmentOutPath();
+        await checkOverwrite(outPath);
+        await cutEncodeSmartPartWrapper({ cutFrom: desiredCutFrom, cutTo, outPath });
+        return outPath;
+      }
+
+      const ext = getOutFileExtension({ isCustomFormatSelected: true, outFormat, filePath });
+
+      const smartCutMainPartOutPath = segmentNeedsSmartCut
+        ? getSuffixedOutPath({ customOutDir, filePath, nameSuffix: `smartcut-segment-copy-${i}${ext}` })
+        : await makeSegmentOutPath();
+
+      const smartCutEncodedPartOutPath = getSuffixedOutPath({ customOutDir, filePath, nameSuffix: `smartcut-segment-encode-${i}${ext}` });
+
+      const smartCutSegmentsToConcat = [smartCutEncodedPartOutPath, smartCutMainPartOutPath];
+
+      if (!segmentNeedsSmartCut) await checkOverwrite(smartCutMainPartOutPath);
+
+      // for smart cut we need to use keyframe cut here, and no avoid_negative_ts
+      await cutSingle({
+        cutFrom: encodeCutTo, cutTo, chaptersPath, outPath: smartCutMainPartOutPath, copyFileStreams: copyFileStreamsFiltered, keyframeCut: true, avoidNegativeTs: false, videoDuration, rotation, allFilesMeta, outFormat, appendFfmpegCommandLog, shortestFlag, ffmpegExperimental, preserveMovData, movFastStart, customTagsByFile, customTagsByStreamId, dispositionByStreamId, videoTimebase, onProgress: onCutProgress,
+      });
+
+      // OK, just return the single cut file (we may need smart cut in other segments though)
+      if (!segmentNeedsSmartCut) return smartCutMainPartOutPath;
+
+      try {
+        const frameDuration = 1 / detectedFps;
+        const encodeCutToSafe = Math.max(desiredCutFrom + frameDuration, encodeCutTo - frameDuration); // Subtract one frame so we don't end up with duplicates when concating, and make sure we don't create a 0 length segment
+
+        await cutEncodeSmartPartWrapper({ cutFrom: desiredCutFrom, cutTo: encodeCutToSafe, outPath: smartCutEncodedPartOutPath });
+
+        // need to re-read streams because indexes may have changed. Using main file as source of streams and metadata
+        const { streams: streamsAfterCut } = await readFileMeta(smartCutMainPartOutPath);
+
+        const outPath = await makeSegmentOutPath();
+        await checkOverwrite(outPath);
+
+        await concatFiles({ paths: smartCutSegmentsToConcat, outDir: outputDir, outPath, metadataFromPath: smartCutMainPartOutPath, outFormat, includeAllStreams: true, streams: streamsAfterCut, ffmpegExperimental, preserveMovData, movFastStart, chapters, preserveMetadataOnMerge, videoTimebase, appendFfmpegCommandLog, onProgress: onConcatProgress });
+        return outPath;
+      } finally {
+        await tryDeleteFiles(smartCutSegmentsToConcat);
+      }
+    }
 
     try {
-      // Keep this similar to cutSingle()
-      const ffmpegArgs = [
-        '-hide_banner',
-        // No progress if we set loglevel warning :(
-        // '-loglevel', 'warning',
+      const outFiles = await pMap(segments, maybeSmartCutSegment, { concurrency: 1 });
 
-        // https://blog.yo1.dog/fix-for-ffmpeg-protocol-not-on-whitelist-error-for-urls/
-        '-f', 'concat', '-safe', '0', '-protocol_whitelist', 'file,pipe', '-i', '-',
-
-        // Add the first file for using its metadata. Can only do this if allStreams (-map 0) is set, or else ffmpeg might output this input instead of the concat
-        ...(preserveMetadataOnMerge && allStreams ? ['-i', paths[0]] : []),
-
-        // Chapters?
-        ...(ffmetadataPath ? ['-f', 'ffmetadata', '-i', ffmetadataPath] : []),
-
-        '-c', 'copy',
-
-        ...(allStreams ? ['-map', '0'] : []),
-
-        // Use the file index 1 for metadata
-        // -map_metadata 0 with concat demuxer doesn't seem to preserve metadata when merging.
-        // Can only do this if allStreams (-map 0) is set
-        ...(preserveMetadataOnMerge && allStreams ? ['-map_metadata', '1'] : []),
-
-        ...getMovFlags({ preserveMovData, movFastStart }),
-
-        // See https://github.com/mifi/lossless-cut/issues/170
-        '-ignore_unknown',
-
-        // https://superuser.com/questions/543589/information-about-ffmpeg-command-line-options
-        ...(ffmpegExperimental ? ['-strict', 'experimental'] : []),
-
-        ...(outFormat ? ['-f', outFormat] : []),
-        '-y', outPath,
-      ];
-
-      console.log('ffmpeg', ffmpegArgs.join(' '));
-
-      // https://superuser.com/questions/787064/filename-quoting-in-ffmpeg-concat
-      const concatTxt = paths.map(file => `file '${join(file).replace(/'/g, "'\\''")}'`).join('\n');
-
-      console.log(concatTxt);
-
-      const ffmpegPath = getFfmpegPath();
-      const process = execa(ffmpegPath, ffmpegArgs);
-
-      handleProgress(process, totalDuration, onProgress);
-
-      stringToStream(concatTxt).pipe(process.stdin);
-
-      const { stdout } = await process;
-      console.log(stdout);
+      return outFiles;
     } finally {
-      if (ffmetadataPath) await fs.unlink(ffmetadataPath).catch((err) => console.error('Failed to delete', ffmetadataPath, err));
+      if (chaptersPath) await tryDeleteFiles([chaptersPath]);
     }
+  }, [concatFiles, cutSingle, filePath, needSmartCut]);
 
-    await optionalTransferTimestamps(paths[0], outPath);
-  }, [optionalTransferTimestamps]);
-
-  const autoMergeSegments = useCallback(async ({ customOutDir, isCustomFormatSelected, outFormat, segmentPaths, ffmpegExperimental, onProgress, preserveMovData, movFastStart, autoDeleteMergedSegments, chapterNames, preserveMetadataOnMerge }) => {
+  const autoConcatCutSegments = useCallback(async ({ customOutDir, isCustomFormatSelected, outFormat, segmentPaths, ffmpegExperimental, onProgress, preserveMovData, movFastStart, autoDeleteMergedSegments, chapterNames, preserveMetadataOnMerge, appendFfmpegCommandLog }) => {
     const ext = getOutFileExtension({ isCustomFormatSelected, outFormat, filePath });
-    const fileName = `cut-merged-${new Date().getTime()}${ext}`;
-    const outPath = getOutPath(customOutDir, filePath, fileName);
+    const outPath = getSuffixedOutPath({ customOutDir, filePath, nameSuffix: `cut-merged-${new Date().getTime()}${ext}` });
     const outDir = getOutDir(customOutDir, filePath);
 
     const chapters = await createChaptersFromSegments({ segmentPaths, chapterNames });
 
-    await mergeFiles({ paths: segmentPaths, outDir, outPath, outFormat, allStreams: true, ffmpegExperimental, onProgress, preserveMovData, movFastStart, chapters, preserveMetadataOnMerge });
-    if (autoDeleteMergedSegments) await pMap(segmentPaths, path => fs.unlink(path), { concurrency: 5 });
-  }, [filePath, mergeFiles]);
+    const metadataFromPath = segmentPaths[0];
+    // need to re-read streams because may have changed
+    const { streams } = await readFileMeta(metadataFromPath);
+    await concatFiles({ paths: segmentPaths, outDir, outPath, metadataFromPath, outFormat, includeAllStreams: true, streams, ffmpegExperimental, onProgress, preserveMovData, movFastStart, chapters, preserveMetadataOnMerge, appendFfmpegCommandLog });
+    if (autoDeleteMergedSegments) await tryDeleteFiles(segmentPaths);
 
-  const html5ify = useCallback(async ({ filePath: specificFilePath, outPath, video, audio, onProgress }) => {
-    console.log('Making HTML5 friendly version', { specificFilePath, outPath, video, audio });
+    return outPath;
+  }, [concatFiles, filePath]);
 
-    let videoArgs;
-    let audioArgs;
-
-    // h264/aac_at: No licensing when using HW encoder (Video/Audio Toolbox on Mac)
-    // https://github.com/mifi/lossless-cut/issues/372#issuecomment-810766512
-
-    const targetHeight = 400;
-
-    switch (video) {
-      case 'hq': {
-        if (isMac) {
-          videoArgs = ['-vf', 'format=yuv420p', '-allow_sw', '1', '-vcodec', 'h264', '-b:v', '15M'];
-        } else {
-          // AV1 is very slow
-          // videoArgs = ['-vf', 'format=yuv420p', '-sws_flags', 'neighbor', '-vcodec', 'libaom-av1', '-crf', '30', '-cpu-used', '8'];
-          // Theora is a bit faster but not that much
-          // videoArgs = ['-vf', '-c:v', 'libtheora', '-qscale:v', '1'];
-          // videoArgs = ['-vf', 'format=yuv420p', '-c:v', 'libvpx-vp9', '-crf', '30', '-b:v', '0', '-row-mt', '1'];
-          // x264 can only be used in GPL projects
-          videoArgs = ['-vf', 'format=yuv420p', '-c:v', 'libx264', '-profile:v', 'high', '-preset:v', 'slow', '-crf', '17'];
-        }
-        break;
-      }
-      case 'lq': {
-        if (isMac) {
-          videoArgs = ['-vf', `scale=-2:${targetHeight},format=yuv420p`, '-allow_sw', '1', '-sws_flags', 'lanczos', '-vcodec', 'h264', '-b:v', '1500k'];
-        } else {
-          // videoArgs = ['-vf', `scale=-2:${targetHeight},format=yuv420p`, '-sws_flags', 'neighbor', '-c:v', 'libtheora', '-qscale:v', '1'];
-          // x264 can only be used in GPL projects
-          videoArgs = ['-vf', `scale=-2:${targetHeight},format=yuv420p`, '-sws_flags', 'neighbor', '-c:v', 'libx264', '-profile:v', 'baseline', '-x264opts', 'level=3.0', '-preset:v', 'ultrafast', '-crf', '28'];
-        }
-        break;
-      }
-      case 'copy': {
-        videoArgs = ['-vcodec', 'copy'];
-        break;
-      }
-      default: {
-        videoArgs = ['-vn'];
-      }
-    }
-
-    switch (audio) {
-      case 'hq': {
-        if (isMac) {
-          audioArgs = ['-acodec', 'aac_at', '-b:a', '192k'];
-        } else {
-          audioArgs = ['-acodec', 'flac'];
-        }
-        break;
-      }
-      case 'lq': {
-        if (isMac) {
-          audioArgs = ['-acodec', 'aac_at', '-ar', '44100', '-ac', '2', '-b:a', '96k'];
-        } else {
-          audioArgs = ['-acodec', 'flac', '-ar', '11025', '-ac', '2'];
-        }
-        break;
-      }
-      case 'copy': {
-        audioArgs = ['-acodec', 'copy'];
-        break;
-      }
-      default: {
-        audioArgs = ['-an'];
-      }
-    }
-
-    const ffmpegArgs = [
-      '-hide_banner',
-
-      '-i', specificFilePath,
-      ...videoArgs,
-      ...audioArgs,
-      '-sn',
-      '-y', outPath,
-    ];
-
-    const duration = await getDuration(specificFilePath);
-    const process = runFfmpeg(ffmpegArgs);
-    if (duration) handleProgress(process, duration, onProgress);
-
-    const { stdout } = await process;
-    console.log(stdout);
-
-    await optionalTransferTimestamps(specificFilePath, outPath);
+  const html5ify = useCallback(async ({ customOutDir, filePath: filePathArg, speed, hasAudio, hasVideo, onProgress }) => {
+    const outPath = getHtml5ifiedPath(customOutDir, filePathArg, speed);
+    await ffmpegHtml5ify({ filePath: filePathArg, outPath, speed, hasAudio, hasVideo, onProgress });
+    await optionalTransferTimestamps(filePathArg, outPath);
+    return outPath;
   }, [optionalTransferTimestamps]);
 
   // This is just used to load something into the player with correct length,
   // so user can seek and then we render frames using ffmpeg
-  const html5ifyDummy = useCallback(async ({ filePath: specificFilePath, outPath, onProgress }) => {
-    console.log('Making HTML5 friendly dummy', { specificFilePath, outPath });
+  const html5ifyDummy = useCallback(async ({ filePath: filePathArg, outPath, onProgress }) => {
+    console.log('Making HTML5 friendly dummy', { filePathArg, outPath });
 
-    const duration = await getDuration(specificFilePath);
+    const duration = await getDuration(filePathArg);
 
     const ffmpegArgs = [
       '-hide_banner',
@@ -393,30 +490,36 @@ function useFfmpegOperations({ filePath, enableTransferTimestamps }) {
     const process = runFfmpeg(ffmpegArgs);
     handleProgress(process, duration, onProgress);
 
-    const { stdout } = await process;
-    console.log(stdout);
+    const result = await process;
+    logStdoutStderr(result);
 
-    await optionalTransferTimestamps(specificFilePath, outPath);
+    await optionalTransferTimestamps(filePathArg, outPath);
   }, [optionalTransferTimestamps]);
 
   // https://stackoverflow.com/questions/34118013/how-to-determine-webm-duration-using-ffprobe
-  const fixInvalidDuration = useCallback(async ({ fileFormat, customOutDir }) => {
+  const fixInvalidDuration = useCallback(async ({ fileFormat, customOutDir, duration, onProgress }) => {
     const ext = getOutFileExtension({ outFormat: fileFormat, filePath });
-    const fileName = `reformatted${ext}`;
-    const outPath = getOutPath(customOutDir, filePath, fileName);
+    const outPath = getSuffixedOutPath({ customOutDir, filePath, nameSuffix: `reformatted${ext}` });
 
     const ffmpegArgs = [
       '-hide_banner',
 
       '-i', filePath,
 
+      // https://github.com/mifi/lossless-cut/issues/1415
+      '-map_metadata', '0',
+      '-map', '0',
+      '-ignore_unknown',
+
       '-c', 'copy',
       '-y', outPath,
     ];
 
-    // TODO progress
-    const { stdout } = await runFfmpeg(ffmpegArgs);
-    console.log(stdout);
+    const process = runFfmpeg(ffmpegArgs);
+    handleProgress(process, duration, onProgress);
+
+    const result = await process;
+    logStdoutStderr(result);
 
     await optionalTransferTimestamps(filePath, outPath);
 
@@ -424,7 +527,7 @@ function useFfmpegOperations({ filePath, enableTransferTimestamps }) {
   }, [filePath, optionalTransferTimestamps]);
 
   return {
-    cutMultiple, mergeFiles, html5ify, html5ifyDummy, fixInvalidDuration, autoMergeSegments,
+    cutMultiple, concatFiles, html5ify, html5ifyDummy, fixInvalidDuration, autoConcatCutSegments,
   };
 }
 
