@@ -1,20 +1,36 @@
 import { useCallback } from 'react';
+import { t } from 'i18next';
 
 import normalizeWheel from './normalizeWheel';
 
-function useTimelineScroll({ wheelSensitivity, invertTimelineScroll, zoomRel, seekRel }) {
+export const keyMap = {
+  ctrl: 'ctrlKey',
+  shift: 'shiftKey',
+  alt: 'altKey',
+  meta: 'metaKey',
+};
+
+export const getModifierKeyNames = () => ({
+  ctrl: t('Ctrl'),
+  shift: t('Shift'),
+  alt: t('Alt'),
+  meta: t('⌘ Cmd / ⊞ Win'),
+});
+
+function useTimelineScroll({ wheelSensitivity, mouseWheelZoomModifierKey, invertTimelineScroll, zoomRel, seekRel }) {
   const onWheel = useCallback((e) => {
     const { pixelX, pixelY } = normalizeWheel(e);
     // console.log({ spinX, spinY, pixelX, pixelY });
 
     const direction = invertTimelineScroll ? 1 : -1;
 
-    if (e.ctrlKey) {
+    const modifierKey = keyMap[mouseWheelZoomModifierKey];
+    if (e[modifierKey]) {
       zoomRel(direction * (pixelY) * wheelSensitivity * 0.4);
     } else {
       seekRel(direction * (pixelX + pixelY) * wheelSensitivity * 0.2);
     }
-  }, [seekRel, zoomRel, wheelSensitivity, invertTimelineScroll]);
+  }, [invertTimelineScroll, mouseWheelZoomModifierKey, zoomRel, wheelSensitivity, seekRel]);
 
   return onWheel;
 }
